refactor(functions): split monolithic index.ts into leaf modules; dedupe shared types
F12: reduce functions/src/index.ts from ~4337 lines to a 115-line barrel of
re-exports; extract logical groups into leaf modules (aiGeneration,
studentIdentity, classlinkRoster, driveArchive, embedProxy,
adminAnalyticsEndpoint) plus shared infra leaves (functionsInit with a guarded
single admin.initializeApp, secrets, shared). The set of 61 exported
identifiers (56 Cloud Functions) is byte-identical, so deploy targets are
unchanged; dependency graph is acyclic.
F18: centralize ClassLinkClass / ClassLinkUser / isSafeEmailForOneRosterFilter
in classlinkShared.ts as a single source instead of per-module copies.
ClassLinkClass mirrors the repo-root type but stays functions-local: importing
root types.ts into the functions build emits a stray root types.js that would
shadow types.ts under Vite resolution.

Implements F12 and F18 from docs/optimize-pass/05-build-infra-monorepo.md.
