+++ functions/src/index.ts
@@ -1,32 +1,59 @@
-import { onCall, onRequest, HttpsError } from 'firebase-functions/v2/https';
-import { setGlobalOptions } from 'firebase-functions/v2';
-import { defineSecret } from 'firebase-functions/params';
-import * as admin from 'firebase-admin';
-import axios, { AxiosError } from 'axios';
-import * as CryptoJS from 'crypto-js';
-import { GoogleGenAI, Content, Type, Schema } from '@google/genai';
-import { randomUUID } from 'node:crypto';
-import { OAuth2Client } from 'google-auth-library';
-import { sanitizePrompt } from './sanitize';
-import { parseGeminiJson } from './parseGeminiJson';
-import { BoundedLruMap } from './utils/boundedLruMap';
-import { readAnalyticsSnapshot } from './adminAnalyticsSnapshot';
-import {
-  ALLOWED_ORIGINS,
-  ONEROSTER_BASE,
-  computeStudentUid,
-  getOAuthHeaders,
-  normalizeEmailDomain,
-  resolveOrgIdForDomain,
-  type ClassLinkStudent,
-} from './classlinkShared';
-import { normalizeQuizCode } from './quizCode';
+/**
+ * Cloud Functions entrypoint — a THIN BARREL of re-exports (F12).
+ *
+ * Every Cloud Function lives in its own leaf module; this file only re-exports
+ * them so `firebase deploy --only functions` and the test suite see the exact
+ * same set of exported identifiers as before the split. Adding a function means
+ * adding a leaf module and a re-export line here — NOT editing a 4000-line file.
+ *
+ * INVARIANT: the set of exported identifiers (and their names) must stay
+ * byte-identical, or Firebase would treat renamed/removed exports as deleted
+ * deploy targets. A barrel `export { x } from './m'` preserves the deployed
+ * function name `x`.
+ *
+ * `admin.initializeApp()` + `setGlobalOptions()` are NOT called here anymore —
+ * each leaf module imports the shared `./functionsInit` side-effect module,
+ * which runs them exactly once (guarded on `admin.apps.length`).
+ */
 
-// Phase 4 — organization invitations + membership write-through.
-// These modules initialize their own `admin.initializeApp()` guarded by
-// `admin.apps.length`, so importing here is safe: the first import (either
-// direction) wins and the others no-op. Re-exporting the callables and the
-// trigger makes them deployable via `firebase deploy --only functions`.
+// ── ClassLink roster (teacher-side OneRoster import) ───────────────────────
+export { getClassLinkRosterV1 } from './classlinkRoster';
+
+// ── AI generation (Gemini): quiz / video-activity / guided-learning / etc. ─
+// Includes the test-only validator + cache-introspection re-exports the test
+// suites import from `./index`.
+export {
+  generateWithAI,
+  generateVideoActivity,
+  transcribeVideoWithGemini,
+  generateGuidedLearning,
+  validateAndBucketVideoQuestions,
+  validateAndBucketQuizQuestions,
+  __resetGenerateWithAICaches,
+  __getCachedAdminStatus,
+  __getGeminiModelConfig,
+} from './aiGeneration';
+
+// ── External-content proxy + iframe embeddability check ────────────────────
+export { fetchExternalProxy, checkUrlCompatibility } from './embedProxy';
+
+// ── Activity Wall photo → Google Drive archive ─────────────────────────────
+export { archiveActivityWallPhoto } from './driveArchive';
+
+// ── Admin analytics HTTP endpoint (snapshot read) ──────────────────────────
+export { adminAnalytics } from './adminAnalyticsEndpoint';
+
+// ── Student identity (ClassLink-via-Google SSO) + PIN→SSO unification ───────
+export {
+  studentLoginV1,
+  getAssignmentPseudonymV1,
+  getStudentClassDirectoryV1,
+  getPseudonymsForAssignmentV1,
+  commitRosterPinIndexV1,
+  pinLoginV1,
+} from './studentIdentity';
+
+// ── Organization invitations + membership write-through (Phase 4) ──────────
 export {
   createOrganizationInvites,
   claimOrganizationInvite,
@@ -59,4254 +86,6 @@ export {
   refreshSpotifyAccessToken,
   revokeSpotifyAuth,
 } from './spotifyOAuth';
-
-setGlobalOptions({ region: 'us-central1' });
-
-const GEMINI_API_KEY = defineSecret('GEMINI_API_KEY');
-const CLASSLINK_CLIENT_ID = defineSecret('CLASSLINK_CLIENT_ID');
-const CLASSLINK_CLIENT_SECRET = defineSecret('CLASSLINK_CLIENT_SECRET');
-const CLASSLINK_TENANT_URL = defineSecret('CLASSLINK_TENANT_URL');
-const STUDENT_PSEUDONYM_HMAC_SECRET = defineSecret(
-  'STUDENT_PSEUDONYM_HMAC_SECRET'
-);
-const GOOGLE_OAUTH_CLIENT_ID = defineSecret('GOOGLE_OAUTH_CLIENT_ID');
-
-admin.initializeApp();
-
-interface ClassLinkUser {
-  sourcedId: string;
-  email: string;
-  givenName: string;
-  familyName: string;
-}
-
-interface ClassLinkClass {
-  sourcedId: string;
-  title: string;
-  classCode?: string;
-}
-
-interface AIData {
-  type:
-    | 'mini-app'
-    | 'poll'
-    | 'dashboard-layout'
-    | 'instructional-routine'
-    | 'ocr'
-    | 'quiz'
-    | 'widget-builder'
-    | 'widget-explainer'
-    | 'blooms-ai'
-    | 'video-activity-recommend';
-  prompt?: string;
-  image?: string; // base64 data
-  /**
-   * Per-type quiz question counts. Only honored for `type === 'quiz'`.
-   * Each entry is the number of questions of that type the teacher
-   * requested. Unspecified types are treated as zero.
-   */
-  typeCounts?: Partial<Record<QuizGenType, number>>;
-}
-
-type QuizGenType = 'MC' | 'FIB' | 'Matching' | 'Ordering';
-
-interface GlobalPermConfig {
-  dailyLimit?: number;
-  dailyLimitEnabled?: boolean;
-}
-
-interface GlobalPermission {
-  enabled: boolean;
-  accessLevel: 'admin' | 'beta' | 'all';
-  betaUsers?: string[];
-  config?: GlobalPermConfig;
-}
-
-const DEFAULT_ADVANCED_MODEL = 'gemini-3-flash-preview';
-const DEFAULT_STANDARD_MODEL = 'gemini-3.1-flash-lite-preview';
-
-/**
- * Validates and normalises a Gemini model name.
- * Returns `undefined` when the supplied value is falsy or fails the pattern
- * check, so callers can fall back to a default.
- */
-function normalizeModelName(raw: unknown): string | undefined {
-  if (!raw || typeof raw !== 'string') return undefined;
-  const trimmed = raw.trim();
-  if (!trimmed) return undefined;
-  if (!/^gemini-[\w.-]+$/.test(trimmed)) return undefined;
-  return trimmed;
-}
-
-/**
- * Splits an array into fixed-size chunks (last chunk may be short).
- * Used to bound fan-out parallelism — both for Firestore `in` queries
- * (10-item limit per query) and for external HTTP fan-outs (ClassLink,
- * etc.) where unbounded `Promise.all` can OOM the function instance or
- * hammer the upstream API.
- */
-function chunk<T>(arr: readonly T[], size: number): T[][] {
-  const chunks: T[][] = [];
-  for (let i = 0; i < arr.length; i += size) {
-    chunks.push(arr.slice(i, i + size));
-  }
-  return chunks;
-}
-
-interface GeminiModelConfig {
-  advancedModel?: string;
-  standardModel?: string;
-}
-
-// Module-scope read caches for `generateWithAI`. Cloud Functions 2nd-gen
-// reuses warm instances, so caching across invocations within a warm
-// instance materially cuts Firestore reads for high-traffic AI features
-// (audit doc, item #3 — was 4–6 reads per call, now 2 transactional reads
-// on warm hits). 5-minute TTL keeps admin demotions and model-config
-// overrides propagating within an acceptable window.
-//
-// NOT cached: `ai_usage/*` counter reads and the in-transaction
-// `global_permissions/*` reads. Those gate rate-limit enforcement and
-// must be transactional to prevent races.
-const READ_CACHE_TTL_MS = 5 * 60 * 1000;
-
-interface ModelConfigCacheEntry {
-  value: {
-    advancedModel: string;
-    standardModel: string;
-    usedFallback: boolean;
-  };
-  cachedAt: number;
-}
-let cachedModelConfig: ModelConfigCacheEntry | null = null;
-
-interface AdminStatusCacheEntry {
-  isAdmin: boolean;
-  cachedAt: number;
-}
-
-// Bound on `cachedAdminStatus` size. A warm Cloud Functions instance that
-// sees many distinct callers across its lifetime would otherwise grow this
-// Map unboundedly. At school-district scale this is firmly in "won't
-// matter" territory (low thousands of admins org-wide), but a hard cap
-// closes the long-tail memory growth path that two independent reviewers
-// flagged on PR #1590. `BoundedLruMap` promotes on read so frequently-hit
-// keys survive eviction pressure from one-off callers — this becomes
-// load-bearing once the same pattern is reused on a higher-cardinality key
-// space (e.g. the student-pseudonym path).
-const ADMIN_STATUS_CACHE_MAX = 500;
-const cachedAdminStatus = new BoundedLruMap<string, AdminStatusCacheEntry>(
-  ADMIN_STATUS_CACHE_MAX
-);
-
-/**
- * Test-only: reset the module-scope read caches so tests can observe a
- * cold-start read sequence. Underscored prefix makes it obvious this is
- * not a production API.
- */
-export function __resetGenerateWithAICaches(): void {
-  cachedModelConfig = null;
-  cachedAdminStatus.clear();
-}
-
-/**
- * Reads the admin-configured model overrides from the `gemini-functions`
- * global permissions document. Returns validated model names (or defaults).
- * Memoized with a 5-minute TTL — see `READ_CACHE_TTL_MS` above.
- *
- * `usedFallback` is `true` when the Firestore read threw — meaning the
- * caller is running with hardcoded defaults rather than whatever overrides
- * an admin may have configured. Plumb this back to the client so admins
- * get a one-time UI signal during Firestore brownouts. Cache hits always
- * return `usedFallback: false` because the catch path deliberately does
- * NOT populate the cache.
- */
-async function getGeminiModelConfig(db: admin.firestore.Firestore): Promise<{
-  advancedModel: string;
-  standardModel: string;
-  usedFallback: boolean;
-}> {
-  const now = Date.now();
-  if (
-    cachedModelConfig &&
-    now - cachedModelConfig.cachedAt < READ_CACHE_TTL_MS
-  ) {
-    return cachedModelConfig.value;
-  }
-  try {
-    const doc = await db
-      .collection('global_permissions')
-      .doc('gemini-functions')
-      .get();
-    const cfg = doc.data()?.config as GeminiModelConfig | undefined;
-    const value = {
-      advancedModel:
-        normalizeModelName(cfg?.advancedModel) ?? DEFAULT_ADVANCED_MODEL,
-      standardModel:
-        normalizeModelName(cfg?.standardModel) ?? DEFAULT_STANDARD_MODEL,
-      usedFallback: false,
-    };
-    cachedModelConfig = { value, cachedAt: now };
-    return value;
-  } catch (error) {
-    console.warn(
-      'Failed to read Gemini model config from Firestore; using defaults.',
-      error
-    );
-    // Do not cache the fallback — a transient Firestore error shouldn't
-    // pin the function to defaults for 5 minutes.
-    return {
-      advancedModel: DEFAULT_ADVANCED_MODEL,
-      standardModel: DEFAULT_STANDARD_MODEL,
-      usedFallback: true,
-    };
-  }
-}
-
-/**
- * Memoized admin lookup. The `admins/{email}` doc rarely changes during a
- * single warm-instance lifetime, so caching for 5 minutes saves one
- * Firestore read per warm `generateWithAI` call. Key is the lowercased
- * email (matches the production read).
- *
- * Trade-offs: TTL bounds both admin demotion lag (cached `true` survives
- * a demotion for up to 5 min — they keep elevated quota briefly) and
- * promotion lag (cached `false` survives a promotion for up to 5 min —
- * they see old limits briefly). Both are acceptable.
- *
- * No try/catch here on purpose. A Firestore failure must throw up to the
- * caller — caching a wrong answer would either grant non-admins admin
- * quotas (false positive) or strip admins of their quotas (false
- * negative), both worse than a temporary 5xx that the client retries.
- */
-async function getCachedAdminStatus(
-  db: admin.firestore.Firestore,
-  emailLower: string
-): Promise<boolean> {
-  const now = Date.now();
-  const cached = cachedAdminStatus.get(emailLower);
-  if (cached) {
-    if (now - cached.cachedAt < READ_CACHE_TTL_MS) {
-      return cached.isAdmin;
-    }
-    // Prune stale entries so a long-lived warm instance that sees many
-    // unique callers doesn't accumulate dead Map entries.
-    cachedAdminStatus.delete(emailLower);
-  }
-  const doc = await db.collection('admins').doc(emailLower).get();
-  const isAdmin = doc.exists;
-  cachedAdminStatus.set(emailLower, { isAdmin, cachedAt: now });
-  return isAdmin;
-}
-
-// Test-only re-exports so the cache contract can be verified without
-// driving through the full `generateWithAI` pipeline (which would require
-// mocking `@google/genai`). The `__` prefix makes the test-only status
-// grep-able.
-export {
-  getCachedAdminStatus as __getCachedAdminStatus,
-  getGeminiModelConfig as __getGeminiModelConfig,
-};
-
-interface ArchiveActivityWallPhotoData {
-  accessToken?: string;
-  sessionId?: string;
-  submissionId?: string;
-  activityId?: string;
-  status?: 'approved' | 'pending';
-}
-
-const DRIVE_API_URL = 'https://www.googleapis.com/drive/v3';
-const UPLOAD_API_URL = 'https://www.googleapis.com/upload/drive/v3';
-const APP_DRIVE_FOLDER = 'SpartBoard';
-
-const getDriveHeaders = (accessToken: string) => ({
-  Authorization: `Bearer ${accessToken}`,
-  'Content-Type': 'application/json',
-});
-
-const listDriveFiles = async (
-  accessToken: string,
-  query: string
-): Promise<Array<{ id: string; name: string }>> => {
-  const url = new URL(`${DRIVE_API_URL}/files`);
-  url.searchParams.set('q', query);
-  url.searchParams.set('fields', 'files(id,name)');
-
-  const response = await fetch(url.toString(), {
-    headers: getDriveHeaders(accessToken),
-  });
-
-  if (!response.ok) {
-    throw new Error(`Failed to list Drive files (${response.status})`);
-  }
-
-  const data = (await response.json()) as {
-    files?: Array<{ id: string; name: string }>;
-  };
-  return data.files ?? [];
-};
-
-const getOrCreateDriveFolder = async (
-  accessToken: string,
-  folderName: string,
-  parentId?: string
-): Promise<string> => {
-  const escapedFolderName = folderName
-    .replace(/\\/g, '\\\\')
-    .replace(/'/g, "\\'");
-  let query = `name = '${escapedFolderName}' and mimeType = 'application/vnd.google-apps.folder' and trashed = false`;
-  if (parentId) {
-    query += ` and '${parentId}' in parents`;
-  }
-
-  const existing = await listDriveFiles(accessToken, query);
-  if (existing[0]?.id) return existing[0].id;
-
-  const response = await fetch(`${DRIVE_API_URL}/files`, {
-    method: 'POST',
-    headers: getDriveHeaders(accessToken),
-    body: JSON.stringify({
-      name: folderName,
-      mimeType: 'application/vnd.google-apps.folder',
-      ...(parentId ? { parents: [parentId] } : {}),
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error(`Failed to create Drive folder ${folderName}`);
-  }
-
-  const folder = (await response.json()) as { id: string };
-  return folder.id;
-};
-
-const getDriveFolderPath = async (
-  accessToken: string,
-  path: string
-): Promise<string> => {
-  const parts = path.split('/').filter(Boolean);
-  let parentId = await getOrCreateDriveFolder(accessToken, APP_DRIVE_FOLDER);
-
-  for (const part of parts) {
-    parentId = await getOrCreateDriveFolder(accessToken, part, parentId);
-  }
-
-  return parentId;
-};
-
-const uploadBlobToDrive = async (
-  accessToken: string,
-  blob: Buffer,
-  mimeType: string,
-  fileName: string,
-  folderPath: string
-): Promise<{ id: string }> => {
-  const folderId = await getDriveFolderPath(accessToken, folderPath);
-
-  const createResponse = await fetch(`${DRIVE_API_URL}/files`, {
-    method: 'POST',
-    headers: getDriveHeaders(accessToken),
-    body: JSON.stringify({
-      name: fileName,
-      parents: [folderId],
-    }),
-  });
-
-  if (!createResponse.ok) {
-    throw new Error('Failed to create file metadata in Drive');
-  }
-
-  const driveFile = (await createResponse.json()) as { id: string };
-
-  const uploadResponse = await fetch(
-    `${UPLOAD_API_URL}/files/${driveFile.id}?uploadType=media`,
-    {
-      method: 'PATCH',
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-        'Content-Type': mimeType,
-      },
-      body: blob,
-    }
-  );
-
-  if (!uploadResponse.ok) {
-    throw new Error('Failed to upload file content to Drive');
-  }
-
-  return driveFile;
-};
-
-const makeDriveFilePublic = async (
-  accessToken: string,
-  fileId: string
-): Promise<void> => {
-  const response = await fetch(`${DRIVE_API_URL}/files/${fileId}/permissions`, {
-    method: 'POST',
-    headers: getDriveHeaders(accessToken),
-    body: JSON.stringify({ role: 'reader', type: 'anyone' }),
-  });
-
-  if (!response.ok) {
-    throw new Error('Failed to share file in Drive');
-  }
-};
-
-export const getClassLinkRosterV1 = onCall(
-  {
-    memory: '256MiB',
-    secrets: [
-      CLASSLINK_CLIENT_ID,
-      CLASSLINK_CLIENT_SECRET,
-      CLASSLINK_TENANT_URL,
-    ],
-    invoker: 'public',
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    const userEmail = request.auth.token.email;
-    if (!userEmail) {
-      throw new HttpsError(
-        'invalid-argument',
-        'User must have an email associated with their account.'
-      );
-    }
-
-    const clientId = CLASSLINK_CLIENT_ID.value();
-    const clientSecret = CLASSLINK_CLIENT_SECRET.value();
-    const tenantUrl = CLASSLINK_TENANT_URL.value();
-
-    if (!clientId || !clientSecret || !tenantUrl) {
-      throw new HttpsError(
-        'internal',
-        'ClassLink configuration is missing on the server.'
-      );
-    }
-
-    const cleanTenantUrl = tenantUrl.replace(/\/$/, '');
-
-    try {
-      if (!isSafeEmailForOneRosterFilter(userEmail)) {
-        return { classes: [], studentsByClass: {} };
-      }
-      const usersBaseUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users`;
-      const userParams = { filter: `email='${userEmail}'` };
-
-      const userHeaders = getOAuthHeaders(
-        usersBaseUrl,
-        userParams,
-        'GET',
-        clientId,
-        clientSecret
-      );
-
-      const userResponse = await axios.get<{ users: ClassLinkUser[] }>(
-        usersBaseUrl,
-        {
-          params: userParams,
-          headers: { ...userHeaders },
-        }
-      );
-
-      const users = userResponse.data.users;
-
-      if (!users || users.length === 0) {
-        return { classes: [], studentsByClass: {} };
-      }
-
-      const teacherSourcedId = users[0].sourcedId;
-
-      const classesUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users/${teacherSourcedId}/classes`;
-      const classesHeaders = getOAuthHeaders(
-        classesUrl,
-        {},
-        'GET',
-        clientId,
-        clientSecret
-      );
-
-      const classesResponse = await axios.get<{ classes: ClassLinkClass[] }>(
-        classesUrl,
-        { headers: { ...classesHeaders } }
-      );
-      const classes = classesResponse.data.classes;
-
-      const studentsByClass: Record<string, ClassLinkStudent[]> = {};
-
-      // Chunk the per-class student lookups so a teacher with 100+ classes
-      // doesn't fire 100+ simultaneous HTTP requests at ClassLink. Audit
-      // item #5. Batch size 15 is the mid-point of the audit's 10-20 range.
-      const CLASSLINK_FANOUT_CHUNK = 15;
-      for (const classBatch of chunk(classes, CLASSLINK_FANOUT_CHUNK)) {
-        await Promise.all(
-          classBatch.map(async (cls: ClassLinkClass) => {
-            const studentsUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/classes/${cls.sourcedId}/students`;
-            const studentsHeaders = getOAuthHeaders(
-              studentsUrl,
-              {},
-              'GET',
-              clientId,
-              clientSecret
-            );
-            try {
-              const studentsResponse = await axios.get<{
-                users: ClassLinkStudent[];
-              }>(studentsUrl, { headers: { ...studentsHeaders } });
-              studentsByClass[cls.sourcedId] =
-                studentsResponse.data.users ?? [];
-            } catch (err) {
-              // Single-class failures aren't fatal (the teacher's other
-              // classes should still load), but they must be visible —
-              // an empty `[]` here without a log makes ClassLink auth
-              // expiry or per-class permission issues silently look like
-              // "this class has no students" to the teacher.
-              console.warn(
-                '[getClassLinkRosterV1] students fetch failed for class',
-                {
-                  classId: cls.sourcedId,
-                  error: err instanceof Error ? err.message : String(err),
-                }
-              );
-              studentsByClass[cls.sourcedId] = [];
-            }
-          })
-        );
-      }
-
-      return {
-        classes,
-        studentsByClass,
-      };
-    } catch (error: unknown) {
-      if (axios.isAxiosError(error)) {
-        const axiosError = error as AxiosError;
-        throw new HttpsError(
-          'internal',
-          `Failed to fetch data from ClassLink: ${axiosError.message}`
-        );
-      }
-      throw new HttpsError('internal', 'Failed to fetch data from ClassLink');
-    }
-  }
-);
-
-export const generateWithAI = onCall(
-  {
-    memory: '512MiB',
-    secrets: [GEMINI_API_KEY],
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = request.data as AIData;
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    const uid = request.auth.uid;
-    const email = request.auth.token.email;
-
-    if (!email) {
-      throw new HttpsError(
-        'invalid-argument',
-        'User must have an email associated with their account.'
-      );
-    }
-
-    const db = admin.firestore();
-
-    // Check if user is an admin (cached for 5 minutes per warm instance).
-    const isAdmin = await getCachedAdminStatus(db, email.toLowerCase());
-
-    // 1. Determine specific feature ID if applicable
-    let specificFeatureId: string | null = null;
-    const genType = String(data?.type || '')
-      .toLowerCase()
-      .trim();
-    if (genType === 'mini-app') specificFeatureId = 'embed-mini-app';
-    if (genType === 'poll') specificFeatureId = 'smart-poll';
-    if (genType === 'quiz') specificFeatureId = 'quiz';
-    if (genType === 'video-activity')
-      specificFeatureId = 'video-activity-audio-transcription';
-    if (genType === 'ocr') specificFeatureId = 'ocr';
-    if (genType === 'guided-learning') specificFeatureId = 'guided-learning';
-    if (genType === 'blooms-ai') specificFeatureId = 'blooms-ai';
-    if (genType === 'video-activity-recommend')
-      specificFeatureId = 'video-activity-recommend';
-    if (genType === 'dashboard-layout') specificFeatureId = 'dashboard-layout';
-    if (genType === 'instructional-routine')
-      specificFeatureId = 'instructional-routine';
-    if (genType === 'widget-builder') specificFeatureId = 'widget-builder';
-    if (genType === 'widget-explainer') specificFeatureId = 'widget-explainer';
-
-    const today = new Date().toISOString().split('T')[0]; // YYYY-MM-DD
-
-    try {
-      await db.runTransaction(async (transaction) => {
-        // --- Read current usage (needed for both tracking and limits) ---
-        const overallUsageRef = db
-          .collection('ai_usage')
-          .doc(`${uid}_${today}`);
-        const overallUsageDoc = await transaction.get(overallUsageRef);
-        const currentOverallUsage = overallUsageDoc.exists
-          ? (overallUsageDoc.data()?.count as number) || 0
-          : 0;
-
-        let specificUsageRef: admin.firestore.DocumentReference | null = null;
-        let currentSpecificUsage = 0;
-
-        if (specificFeatureId) {
-          specificUsageRef = db
-            .collection('ai_usage')
-            .doc(`${uid}_${specificFeatureId}_${today}`);
-          const specUsageDoc = await transaction.get(specificUsageRef);
-          currentSpecificUsage = specUsageDoc.exists
-            ? (specUsageDoc.data()?.count as number) || 0
-            : 0;
-        }
-
-        // --- Rate-limit checks (non-admin only) ---
-        if (!isAdmin) {
-          const globalPermDoc = await transaction.get(
-            db.collection('global_permissions').doc('gemini-functions')
-          );
-          const globalPerm = globalPermDoc.exists
-            ? (globalPermDoc.data() as GlobalPermission)
-            : null;
-
-          if (globalPerm && !globalPerm.enabled) {
-            throw new HttpsError(
-              'permission-denied',
-              'Gemini functions are currently disabled by an administrator.'
-            );
-          }
-
-          if (globalPerm) {
-            const { accessLevel, betaUsers = [] } = globalPerm;
-            if (accessLevel === 'admin') {
-              throw new HttpsError(
-                'permission-denied',
-                'Gemini functions are currently restricted to administrators.'
-              );
-            }
-            if (
-              accessLevel === 'beta' &&
-              !betaUsers.includes(email.toLowerCase())
-            ) {
-              throw new HttpsError(
-                'permission-denied',
-                'You do not have access to Gemini beta functions.'
-              );
-            }
-          }
-
-          const overallLimitEnabled =
-            globalPerm?.config?.dailyLimitEnabled !== false;
-          const overallLimit = globalPerm?.config?.dailyLimit ?? 20;
-
-          if (overallLimitEnabled && currentOverallUsage >= overallLimit) {
-            throw new HttpsError(
-              'resource-exhausted',
-              `Daily AI usage limit reached (${overallLimit} generations). Please try again tomorrow.`
-            );
-          }
-
-          // --- Check Specific Feature Limit ---
-          if (specificFeatureId) {
-            const specPermDoc = await transaction.get(
-              db.collection('global_permissions').doc(specificFeatureId)
-            );
-            if (specPermDoc.exists) {
-              const specPerm = specPermDoc.data() as GlobalPermission;
-              const specLimitEnabled =
-                specPerm.config?.dailyLimitEnabled !== false;
-              const specificLimit = specPerm.config?.dailyLimit ?? 20;
-
-              if (specLimitEnabled && currentSpecificUsage >= specificLimit) {
-                throw new HttpsError(
-                  'resource-exhausted',
-                  `Daily limit for ${specificFeatureId} reached (${specificLimit} per day). Please try again tomorrow.`
-                );
-              }
-            }
-          }
-        }
-
-        // --- Increment usage counters (all users including admins) ---
-        transaction.set(
-          overallUsageRef,
-          {
-            count: currentOverallUsage + 1,
-            email: email,
-            lastUsed: admin.firestore.FieldValue.serverTimestamp(),
-          },
-          { merge: true }
-        );
-
-        if (specificUsageRef) {
-          transaction.set(
-            specificUsageRef,
-            {
-              count: currentSpecificUsage + 1,
-              email: email,
-              lastUsed: admin.firestore.FieldValue.serverTimestamp(),
-            },
-            { merge: true }
-          );
-        }
-      });
-    } catch (error) {
-      if (error instanceof HttpsError) {
-        throw error;
-      }
-      console.error('Usage tracking error:', error);
-      // Don't block AI generation if tracking fails
-      console.warn('AI usage tracking failed, proceeding with generation.');
-    }
-
-    // Read model config from Firestore (for both admins and non-admins)
-    const geminiConfig = await getGeminiModelConfig(db);
-
-    const apiKey = GEMINI_API_KEY.value();
-    if (!apiKey) {
-      console.error('CRITICAL: GEMINI_API_KEY is missing');
-      throw new HttpsError(
-        'internal',
-        'Gemini API Key is missing on the server.'
-      );
-    }
-
-    try {
-      const genType = String(data?.type || '')
-        .toLowerCase()
-        .trim();
-
-      const ai = new GoogleGenAI({ apiKey });
-
-      // Input size guards
-      if (data?.prompt && String(data.prompt).length > 10000) {
-        throw new HttpsError(
-          'invalid-argument',
-          'Prompt exceeds maximum length of 10,000 characters.'
-        );
-      }
-      if (data?.image && String(data.image).length > 5 * 1024 * 1024) {
-        throw new HttpsError(
-          'invalid-argument',
-          'Image exceeds maximum size of 5MB.'
-        );
-      }
-
-      const sanitizedUserInput = sanitizePrompt(data?.prompt);
-
-      const promptMap: Record<
-        string,
-        () => { systemPrompt: string; userPrompt: string }
-      > = {
-        'mini-app': () => ({
-          systemPrompt: `
-You are an expert frontend developer. Create a single-file HTML/JS mini-app based on the user's request provided within <user_request> tags.
-
-Output requirements:
-1. Single file — all CSS and JS embedded inline. No external scripts except Tailwind CDN.
-2. Use Tailwind CDN (<script src="https://cdn.tailwindcss.com"></script>).
-3. Return JSON only: { "title": "...", "html": "..." }
-
-Submission protocol (MANDATORY — every app must implement this exactly):
-
-A. Render a visible <button data-spart-submit> somewhere obvious (usually near the bottom of the main content). Its label should be meaningful for the activity ("Submit", "Done", "Turn In", etc.). Style it with Tailwind so it looks like a primary action.
-
-B. On load, register a persistent listener for init messages from the parent and show/hide the submit button based on the latest message. The parent may re-send SPART_MINIAPP_INIT at any time (e.g. the teacher flips the Submissions toggle mid-session), so the handler MUST stay registered and re-apply state on every message — do NOT use { once: true } and do NOT remove the listener after the first message:
-
-   window.addEventListener('message', (event) => {
-     if (event.data && event.data.type === 'SPART_MINIAPP_INIT') {
-       const enabled = event.data.payload && event.data.payload.submissionsEnabled === true;
-       document.querySelectorAll('[data-spart-submit]').forEach((el) => {
-         el.style.display = enabled ? '' : 'none';
-       });
-     }
-   });
-
-C. When (and ONLY when) the user clicks the submit button, post the result to the parent. Use event delegation on window with closest('[data-spart-submit]') so the handler survives any DOM re-renders and catches multiple submit buttons:
-
-   window.addEventListener('click', (event) => {
-     const btn = event.target.closest && event.target.closest('[data-spart-submit]');
-     if (!btn) return;
-     window.parent.postMessage({
-       type: 'SPART_MINIAPP_RESULT',
-       payload: { /* whatever data the activity produced — object only, no PII */ }
-     }, '*');
-   });
-
-Do NOT auto-submit on every input, key press, or timer tick. The parent treats each SPART_MINIAPP_RESULT message as a student submission, so fire it exactly once per Submit click. The payload must be a plain object (not a scalar or array).
-
-Worked example (flashcards app with a "Done" button):
-
-<!doctype html>
-<html>
-<head>
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="p-6 font-sans">
-  <div id="card" class="text-2xl font-bold text-center p-8 bg-white rounded-xl shadow"></div>
-  <div class="mt-4 flex gap-2 justify-center">
-    <button id="flip" class="px-4 py-2 bg-slate-200 rounded">Flip</button>
-    <button id="next" class="px-4 py-2 bg-slate-200 rounded">Next</button>
-  </div>
-  <div class="mt-6 text-center">
-    <button data-spart-submit class="px-6 py-3 bg-indigo-600 text-white font-bold rounded-xl">Finish</button>
-  </div>
-  <script>
-    const cards = [{ q: '2+2', a: '4' }, { q: '3x3', a: '9' }];
-    let i = 0, showA = false;
-    const el = document.getElementById('card');
-    const render = () => { el.textContent = showA ? cards[i].a : cards[i].q; };
-    document.getElementById('flip').onclick = () => { showA = !showA; render(); };
-    document.getElementById('next').onclick = () => { i = (i+1)%cards.length; showA=false; render(); };
-    render();
-
-    window.addEventListener('message', (e) => {
-      if (e.data && e.data.type === 'SPART_MINIAPP_INIT') {
-        const enabled = e.data.payload && e.data.payload.submissionsEnabled === true;
-        document.querySelectorAll('[data-spart-submit]').forEach((b) => {
-          b.style.display = enabled ? '' : 'none';
-        });
-      }
-    });
-
-    window.addEventListener('click', (e) => {
-      const btn = e.target.closest && e.target.closest('[data-spart-submit]');
-      if (!btn) return;
-      window.parent.postMessage({
-        type: 'SPART_MINIAPP_RESULT',
-        payload: { reviewed: cards.length, completedAt: Date.now() }
-      }, '*');
-    });
-  </script>
-</body>
-</html>
-        `,
-          userPrompt: `User Request: <user_request>${sanitizedUserInput}</user_request>`,
-        }),
-        poll: () => ({
-          systemPrompt: `
-          You are an expert teacher. Create a 4-option multiple choice poll JSON based on the topic provided within <topic> tags:
-          { "question": "...", "options": ["...", "...", "...", "..."] }
-        `,
-          userPrompt: `Topic: <topic>${sanitizedUserInput}</topic>`,
-        }),
-        'dashboard-layout': () => ({
-          systemPrompt: `
-          You are an expert instructional designer and classroom space planner. Based on the user's lesson description provided within <lesson_description> tags, suggest a set of interactive widgets and arrange them on a 12x12 grid (columns 0-11, rows 0-11).
-
-          Available Widgets (use EXACT type strings):
-          - clock: Digital/analog clock
-          - time-tool: Timer/Stopwatch
-          - traffic: Traffic light for behavior/status
-          - text: Simple sticky note/text area
-          - checklist: To-do list
-          - random: Student/item picker
-          - dice: Random dice roller
-          - sound: Noise level meter
-          - drawing: Sketchpad
-          - qr: QR code generator
-          - embed: Website embedder
-          - poll: Multiple choice poll
-          - quiz: Interactive classroom quizzes
-          - webcam: Live camera feed with OCR capabilities
-          - scoreboard: Point tracker
-          - expectations: Classroom expectations icons
-          - weather: Local weather display
-          - schedule: Daily class schedule
-          - calendar: Monthly events
-          - lunchCount: Student meal tracker
-          - classes: Class/Period selector
-          - instructionalRoutines: Library of teaching strategies (e.g. Think-Pair-Share)
-          - materials: Visual list of required student supplies
-          - stickers: Reward/decorative stickers
-          - seating-chart: Classroom layout manager
-          - catalyst: Instructional warm-ups/activities
-
-          Spatial Grid Rules (12x12):
-          1. Total grid width is 12 columns (0-11). Total grid height is 12 rows (0-11).
-          2. Avoid overlapping widgets.
-          3. For every widget, you MUST provide a gridConfig object with col, row, colSpan, and rowSpan.
-          4. Large, primary widgets (like Quizzes, Scoreboards, or Whiteboards) should have large spans (e.g., colSpan: 8, rowSpan: 8) and be placed centrally.
-          5. Utility widgets (like Timers, Traffic Lights, Dice) should be placed on the edges with smaller spans (e.g., colSpan: 2, rowSpan: 3).
-
-          Requirements:
-          1. Select 3-6 most relevant widgets for the activity.
-          2. Return JSON: { "widgets": [{ "type": "...", "config": {}, "gridConfig": { "col": 0, "row": 0, "colSpan": 4, "rowSpan": 4 } }] }
-          3. 'config' should be an empty object {} unless you are setting a specific property known to that widget (like 'question' for 'poll').
-
-          Example Payload for a 'Review Game':
-          {
-            "widgets": [
-              {
-                "type": "scoreboard",
-                "config": {},
-                "gridConfig": { "col": 1, "row": 0, "colSpan": 10, "rowSpan": 3 }
-              },
-              {
-                "type": "poll",
-                "config": { "question": "Who won the war of 1812?" },
-                "gridConfig": { "col": 1, "row": 3, "colSpan": 7, "rowSpan": 8 }
-              },
-              {
-                "type": "random",
-                "config": { "visualStyle": "wheel" },
-                "gridConfig": { "col": 8, "row": 3, "colSpan": 3, "rowSpan": 8 }
-              }
-            ]
-          }
-        `,
-          userPrompt: `Lesson/Activity Description: <lesson_description>${sanitizedUserInput}</lesson_description>`,
-        }),
-        'instructional-routine': () => ({
-          systemPrompt: `
-          You are an expert instructional designer. Create a classroom instructional routine based on the user's description provided within <description> tags.
-
-          Return JSON:
-          {
-            "name": "Routine Name",
-            "grades": "Grade Range (e.g. K-5, 6-12)",
-            "icon": "Lucide Icon Name (e.g. Brain, Users, Zap)",
-            "color": "Color Name (blue, indigo, violet, purple, fuchsia, pink, rose, red, orange, amber, yellow, lime, green, emerald, teal, cyan, sky, slate, zinc, stone, neutral)",
-            "steps": [
-              {
-                "text": "Step instruction...",
-                "icon": "Lucide Icon Name",
-                "color": "Color Name",
-                "label": "Short Label (e.g. Pair, Share)"
-              }
-            ]
-          }
-        `,
-          userPrompt: `Description: <description>${sanitizedUserInput}</description>`,
-        }),
-        ocr: () => ({
-          systemPrompt: `
-          You are an expert at extracting text from images (OCR).
-          Analyze the provided image and extract all readable text accurately.
-          Maintain the structure as best as possible.
-          If there are multiple paragraphs, separate them with double newlines.
-          Return JSON: { "text": "extracted text here" }
-        `,
-          userPrompt: 'Extract text from this image.',
-        }),
-        quiz: () => {
-          const { counts: quizCounts, total: quizTotal } =
-            normalizeTypeCounts<QuizGenType>(
-              QUIZ_QUESTION_TYPES,
-              data?.typeCounts
-            );
-          // Back-compat: if the client didn't send `typeCounts`, fall back
-          // to a sensible default (5 MC) so older clients keep working
-          // until they redeploy with the new payload.
-          if (quizTotal <= 0) {
-            quizCounts.MC = 5;
-          }
-          const effectiveTotal =
-            quizTotal > 0
-              ? quizTotal
-              : QUIZ_QUESTION_TYPES.reduce((sum, t) => sum + quizCounts[t], 0);
-          return {
-            systemPrompt: buildQuizPrompt(
-              quizCounts,
-              effectiveTotal,
-              bufferedRequestCount(effectiveTotal),
-              sanitizedUserInput
-            ),
-            userPrompt: '',
-          };
-        },
-        'widget-builder': () => ({
-          systemPrompt: `
-          You are an expert frontend developer creating classroom widgets. Create a complete self-contained HTML widget for a classroom dashboard.
-          Requirements:
-          1. Use vanilla HTML/CSS/JS only (no external libraries except optional Tailwind CDN).
-          2. Use a dark background (#1e293b) with light text.
-          3. Include all styles inline in the HTML file.
-          4. The widget must work in a sandboxed iframe.
-          5. Make buttons and interactive elements large enough for tablet use.
-          6. Output ONLY the complete HTML code, nothing else - no explanations, no markdown.
-          `,
-          userPrompt: `Create a widget that: <user_request>${sanitizedUserInput}</user_request>`,
-        }),
-        'widget-explainer': () => ({
-          systemPrompt: `
-          You are a classroom teacher assistant. Explain what an HTML widget does in 1-2 plain sentences.
-          Use simple language without code jargon.
-          Output ONLY the explanation, nothing else.
-          `,
-          userPrompt: sanitizedUserInput,
-        }),
-        'blooms-ai': () => ({
-          systemPrompt: `
-          You are an expert instructional designer specializing in Bloom's Taxonomy.
-          Generate clear, practical, classroom-ready content for the requested cognitive level and topic.
-          Format your response as a readable bulleted list using plain text (not markdown).
-          Keep each item concise (one sentence). Output ONLY the list, no preamble or closing.
-          `,
-          userPrompt: sanitizedUserInput,
-        }),
-        'video-activity-recommend': () => ({
-          systemPrompt: `
-You are an expert classroom instructional designer recommending YouTube videos for K-12 teachers building Video Activity assignments. The teacher will provide a topic, learning objective, or grade-level description in the <topic> tags. Recommend a SINGLE high-quality, classroom-appropriate YouTube video that fits the topic.
-
-Hard constraints:
-1. The video MUST be on YouTube (not Vimeo, TED.com, etc.) and have a stable 11-character video id.
-2. Prefer videos under 15 minutes — shorter is better for in-class video activities.
-3. Prefer educator-aligned channels (Crash Course, SciShow Kids, Khan Academy, TED-Ed, National Geographic, etc.) over random uploads.
-4. Reject anything with mature content, ads-heavy intros, or low production quality.
-5. If you cannot confidently recommend a real, currently-live YouTube video that fits the topic, return { "videoId": "", "title": "", "rationale": "..." } with an empty videoId — do NOT hallucinate ids.
-
-Output JSON ONLY in this exact shape:
-{
-  "videoId": "11_char_id",
-  "title": "Video title as you remember it",
-  "rationale": "One sentence explaining why this fits the topic and grade level."
-}
-          `,
-          userPrompt: `Topic: <topic>${sanitizedUserInput}</topic>`,
-        }),
-      };
-
-      const promptDataFn = promptMap[genType];
-
-      if (!promptDataFn) {
-        const debugData = {
-          receivedType: data?.type,
-          typeOfReceivedType: typeof data?.type,
-          genType,
-          keys: Object.keys(data || {}),
-        };
-        console.error(
-          'CRITICAL: Invalid generation type encountered:',
-          JSON.stringify(debugData)
-        );
-        throw new HttpsError(
-          'invalid-argument',
-          `V3 ERROR: Invalid generation type: "${data?.type}". Received keys: ${Object.keys(data || {}).join(', ')}`
-        );
-      }
-
-      const { systemPrompt, userPrompt } = promptDataFn();
-
-      const contents: Content[] = [
-        {
-          role: 'user',
-          parts: [{ text: systemPrompt + '\n\n' + userPrompt }],
-        },
-      ];
-
-      // Add image if provided (for OCR or multi-modal prompts)
-      if (data.image && contents[0] && contents[0].parts) {
-        // Strip data:image/png;base64, prefix if present
-        const base64Data = data.image.includes(',')
-          ? data.image.split(',')[1]
-          : data.image;
-
-        contents[0].parts.push({
-          inlineData: {
-            mimeType: 'image/png',
-            data: base64Data,
-          },
-        });
-      }
-
-      // Use higher complexity model for code generation, and lite for OCR and simple JSON tasks
-      // Model names are admin-configurable via global_permissions/gemini-functions
-      const model =
-        genType === 'mini-app' || genType === 'widget-builder'
-          ? geminiConfig.advancedModel
-          : geminiConfig.standardModel;
-
-      // Quiz alone uses structured-output (responseSchema) — it has a fixed
-      // shape we want to enforce. Other generators stay on plain JSON mode
-      // because they have looser, type-specific shapes.
-      const responseSchema =
-        genType === 'quiz' ? buildQuizResponseSchema() : undefined;
-
-      const result = await ai.models.generateContent({
-        model,
-        contents,
-        config: {
-          // widget-builder, widget-explainer, and blooms-ai return plain text; all other types return JSON
-          responseMimeType:
-            genType === 'widget-builder' ||
-            genType === 'widget-explainer' ||
-            genType === 'blooms-ai'
-              ? 'text/plain'
-              : 'application/json',
-          ...(responseSchema ? { responseSchema } : {}),
-        },
-      });
-
-      const text = result.text;
-
-      if (!text) {
-        throw new Error('Empty response from AI');
-      }
-
-      // blooms-ai returns plain text — wrap in { text } for the generic callAI client
-      if (genType === 'blooms-ai') {
-        return { text, _modelConfigUsedFallback: geminiConfig.usedFallback };
-      }
-
-      // widget-builder and widget-explainer return plain text — wrap in { result } for the client
-      if (genType === 'widget-builder' || genType === 'widget-explainer') {
-        return {
-          result: text,
-          _modelConfigUsedFallback: geminiConfig.usedFallback,
-        };
-      }
-
-      const parsed = parseGeminiJson<Record<string, unknown>>(text);
-
-      // Quiz: post-validate and trim per-type quotas. Without this, Gemini
-      // can return a shape that doesn't match the per-type field rules
-      // documented in types.ts (e.g., a "Matching" question with the wrong
-      // pipe encoding) and the editor would silently render garbage.
-      if (genType === 'quiz') {
-        const { counts: quizCounts, total: quizTotal } =
-          normalizeTypeCounts<QuizGenType>(
-            QUIZ_QUESTION_TYPES,
-            data?.typeCounts
-          );
-        if (quizTotal <= 0) quizCounts.MC = 5;
-        const validatedQuestions = validateAndBucketQuizQuestions(
-          parsed.questions,
-          quizCounts
-        );
-        if (validatedQuestions.length === 0) {
-          throw new HttpsError(
-            'internal',
-            'The AI did not produce any usable questions. Try adjusting the topic or the question mix.'
-          );
-        }
-        return {
-          title:
-            typeof parsed.title === 'string' && parsed.title.trim().length > 0
-              ? parsed.title
-              : 'AI-generated quiz',
-          questions: validatedQuestions,
-          _modelConfigUsedFallback: geminiConfig.usedFallback,
-        };
-      }
-
-      // Annotate JSON responses with the fallback flag so the client can
-      // surface a one-time admin notice. Underscore-prefixed to make the
-      // marker obvious and avoid colliding with any field Gemini returns.
-      return { ...parsed, _modelConfigUsedFallback: geminiConfig.usedFallback };
-    } catch (error: unknown) {
-      console.error('AI Generation Error Details:', error);
-
-      // If it's already an HttpsError, just re-throw it
-      if (
-        error &&
-        typeof error === 'object' &&
-        'code' in error &&
-        'message' in error
-      ) {
-        throw error;
-      }
-
-      const detail = error instanceof Error ? error.message : 'unknown error';
-      throw new HttpsError('internal', `AI generation failed: ${detail}`);
-    }
-  }
-);
-
-export const fetchExternalProxy = onCall(
-  {
-    memory: '256MiB',
-    timeoutSeconds: 30,
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = request.data as { url: string };
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    try {
-      const parsedUrl = new URL(data.url);
-      if (
-        parsedUrl.protocol !== 'https:' ||
-        (parsedUrl.hostname !== 'api.openweathermap.org' &&
-          parsedUrl.hostname !== 'owc.enterprise.earthnetworks.com' &&
-          parsedUrl.hostname !== 'orono.api.nutrislice.com')
-      ) {
-        throw new Error('Invalid host or protocol');
-      }
-    } catch {
-      throw new HttpsError(
-        'invalid-argument',
-        'Invalid proxy URL. Only https://api.openweathermap.org, https://owc.enterprise.earthnetworks.com, and https://orono.api.nutrislice.com are allowed.'
-      );
-    }
-
-    // 1 MB cap on upstream responses. Audit item #6 — without this a
-    // misbehaving upstream could buffer an arbitrary-sized body into the
-    // 128MiB function instance and OOM us. We pass both maxContentLength
-    // and maxBodyLength so the limit is enforced regardless of whether
-    // upstream advertises Content-Length or streams chunked.
-    const MAX_RESPONSE_BYTES = 1_048_576;
-
-    try {
-      const response = await axios.get<unknown>(data.url, {
-        maxContentLength: MAX_RESPONSE_BYTES,
-        maxBodyLength: MAX_RESPONSE_BYTES,
-        // SSRF guard: the allowlist check above only validates the
-        // initial URL. Axios follows up to 5 redirects by default, so
-        // an allowlisted host could 302 to an arbitrary off-allowlist
-        // host and we'd happily proxy the response. Disabling redirects
-        // forces 3xx to surface as a request error and keeps the
-        // allowlist load-bearing.
-        maxRedirects: 0,
-      });
-      return response.data;
-    } catch (error: unknown) {
-      console.error('External Proxy Error:', error);
-      // Translate axios's size-limit message into an explicit
-      // resource-exhausted HttpsError so the client can distinguish "the
-      // upstream is too chatty" from a generic network blip. Gating on
-      // `axios.isAxiosError` matches the pattern used elsewhere in this
-      // file and avoids matching on unrelated errors that happen to
-      // contain "maxContentLength" in their message.
-      if (
-        axios.isAxiosError(error) &&
-        /maxContentLength|maxBodyLength/i.test(error.message ?? '')
-      ) {
-        throw new HttpsError(
-          'resource-exhausted',
-          `Upstream response exceeded the ${MAX_RESPONSE_BYTES / 1024} KB proxy limit.`
-        );
-      }
-      const msg =
-        error instanceof Error ? error.message : 'External fetch failed';
-      throw new HttpsError('internal', msg);
-    }
-  }
-);
-
-export const archiveActivityWallPhoto = onCall(
-  {
-    memory: '512MiB',
-    timeoutSeconds: 120,
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = request.data as ArchiveActivityWallPhotoData;
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    const accessToken = data.accessToken?.trim();
-    const sessionId = data.sessionId?.trim();
-    const submissionId = data.submissionId?.trim();
-    const activityId = data.activityId?.trim();
-    const status = data.status === 'pending' ? 'pending' : 'approved';
-
-    if (!accessToken || !sessionId || !submissionId || !activityId) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Missing required archive parameters.'
-      );
-    }
-
-    if (!sessionId.startsWith(`${request.auth.uid}_`)) {
-      throw new HttpsError(
-        'permission-denied',
-        'You can only archive your own Activity Wall submissions.'
-      );
-    }
-
-    const submissionRef = admin
-      .firestore()
-      .collection('activity_wall_sessions')
-      .doc(sessionId)
-      .collection('submissions')
-      .doc(submissionId);
-
-    await submissionRef.set(
-      {
-        status,
-        archiveStatus: 'syncing',
-        archiveStartedAt: Date.now(),
-        archiveError: admin.firestore.FieldValue.delete(),
-      },
-      { merge: true }
-    );
-
-    try {
-      const submissionSnap = await submissionRef.get();
-      if (!submissionSnap.exists) {
-        throw new Error('Submission not found');
-      }
-
-      const submission = submissionSnap.data() as {
-        storagePath?: unknown;
-      };
-      const storagePath =
-        typeof submission.storagePath === 'string'
-          ? submission.storagePath
-          : null;
-
-      if (!storagePath) {
-        throw new Error('Missing Firebase storage path for photo submission');
-      }
-
-      const bucket = admin.storage().bucket();
-      const file = bucket.file(storagePath);
-
-      // Size guard — `file.download()` buffers the entire object into the
-      // 512MiB function instance's memory, so a misbehaving client could
-      // OOM us with a giant photo. Audit doc item #4. Storage metadata
-      // returns `size` as a string in some SDK versions, hence the coerce.
-      // Fail closed on missing/non-numeric size — `NaN > limit` is `false`,
-      // which would bypass the guard and reopen the OOM path.
-      const [metadata] = await file.getMetadata();
-      const sizeBytes = Number(metadata.size);
-      const MAX_PHOTO_BYTES = 50 * 1024 * 1024;
-      if (!Number.isFinite(sizeBytes) || sizeBytes > MAX_PHOTO_BYTES) {
-        throw new HttpsError(
-          'invalid-argument',
-          Number.isFinite(sizeBytes)
-            ? `Photo exceeds the 50 MB archive limit (${Math.round(sizeBytes / 1024 / 1024)} MB).`
-            : 'Photo size unknown; cannot safely archive.'
-        );
-      }
-
-      const [fileBuffer] = await file.download();
-      const mimeType = metadata.contentType || 'image/jpeg';
-      const extension =
-        mimeType === 'image/png'
-          ? 'png'
-          : mimeType === 'image/gif'
-            ? 'gif'
-            : mimeType === 'image/webp'
-              ? 'webp'
-              : 'jpg';
-
-      const driveFile = await uploadBlobToDrive(
-        accessToken,
-        fileBuffer,
-        mimeType,
-        `${submissionId}.${extension}`,
-        `Activity Wall/${activityId}`
-      );
-      await makeDriveFilePublic(accessToken, driveFile.id);
-
-      const driveUrl = `https://lh3.googleusercontent.com/d/${driveFile.id}`;
-
-      await submissionRef.set(
-        {
-          content: driveUrl,
-          status,
-          archiveStatus: 'archived',
-          archiveStartedAt: admin.firestore.FieldValue.delete(),
-          driveFileId: driveFile.id,
-          archivedAt: Date.now(),
-          storagePath: admin.firestore.FieldValue.delete(),
-          archiveError: admin.firestore.FieldValue.delete(),
-        },
-        { merge: true }
-      );
-
-      await file.delete({ ignoreNotFound: true });
-
-      return {
-        archiveStatus: 'archived',
-        driveFileId: driveFile.id,
-        driveUrl,
-      };
-    } catch (error: unknown) {
-      const message =
-        error instanceof Error ? error.message : 'Drive archive failed';
-
-      await submissionRef.set(
-        {
-          status,
-          archiveStatus: 'failed',
-          archiveStartedAt: admin.firestore.FieldValue.delete(),
-          archiveError: message.slice(0, 180),
-        },
-        { merge: true }
-      );
-
-      // Preserve HttpsError codes so the client can distinguish
-      // user-actionable failures (e.g. `invalid-argument` from the size
-      // guard) from genuine server errors. Wrapping every error as
-      // `internal` would tell a teacher whose photo is too large that
-      // SpartBoard is broken when the real fix is to shrink the photo.
-      if (error instanceof HttpsError) {
-        throw error;
-      }
-      throw new HttpsError('internal', message);
-    }
-  }
-);
-
-export const checkUrlCompatibility = onCall(
-  {
-    memory: '256MiB',
-    timeoutSeconds: 20,
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = request.data as { url: string };
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    // Validate URL to prevent SSRF
-    let parsedUrl: URL;
-    try {
-      parsedUrl = new URL(data.url);
-    } catch {
-      throw new HttpsError('invalid-argument', 'Invalid URL provided.');
-    }
-
-    if (parsedUrl.protocol !== 'https:') {
-      throw new HttpsError('invalid-argument', 'Only HTTPS URLs are allowed.');
-    }
-
-    const hostname = parsedUrl.hostname.toLowerCase();
-    // Block private/reserved IP ranges and metadata endpoints.
-    //
-    // IPv4 patterns cover RFC 1918 and link-local ranges; IPv6 patterns cover:
-    //   - Any ::-prefixed address — loopback ::1, unspecified ::,
-    //     IPv4-mapped ::ffff:127.0.0.1, IPv4-compatible ::127.0.0.1.
-    //     Globally routable IPv6 lives in 2000::/3, so :: is always reserved.
-    //   - ULA  fc00::/7  (fc** and fd** prefixes — RFC 4193 private range)
-    //   - Link-local  fe80::/10 (fe80 through febf — never routes globally)
-    //   - Site-local  fec0::/10 (fec0 through feff — deprecated by RFC 3879
-    //     but still private/non-routable, so blocked to be safe)
-    //
-    // Regex note: Node wraps IPv6 hostnames in brackets, e.g. `[::1]`.
-    // The patterns below match the bracketed form as returned by URL.hostname.
-    const blockedPatterns = [
-      /^localhost$/,
-      /^127\./,
-      /^10\./,
-      /^172\.(1[6-9]|2[0-9]|3[01])\./,
-      /^192\.168\./,
-      /^169\.254\./,
-      /^0\./,
-      /^metadata\./,
-      /metadata\.google\.internal/,
-      // Any IPv6 address starting with :: — loopback (::1), unspecified (::),
-      // IPv4-mapped (::ffff:127.0.0.1) and IPv4-compatible (::127.0.0.1).
-      // Globally routable IPv6 is allocated from 2000::/3, so every ::-prefixed
-      // address is reserved, local, or private and must be blocked.
-      /^\[::/,
-      // ULA (fc00::/7 — fc** and fd** prefixes)
-      /^\[f[cd]/,
-      // Link-local (fe80::/10 — fe8x through febx) and site-local
-      // (fec0::/10 — fecx through fefx, deprecated by RFC 3879 but private).
-      /^\[fe[89a-f]/,
-    ];
-    if (blockedPatterns.some((pattern) => pattern.test(hostname))) {
-      throw new HttpsError(
-        'invalid-argument',
-        'URLs pointing to private or reserved IP ranges are not allowed.'
-      );
-    }
-
-    try {
-      const response = await axios.head(data.url, {
-        timeout: 10000,
-        // SSRF guard: the blocklist check above only validates the initial URL.
-        // Without this, axios would follow up to 5 redirects by default, so a
-        // public host could 302 to a private/internal IP (e.g. the GCP metadata
-        // endpoint at 169.254.169.254) and bypass the check entirely — the same
-        // vulnerability that was already fixed in fetchExternalProxy.
-        maxRedirects: 0,
-        headers: {
-          'User-Agent':
-            'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36',
-        },
-      });
-
-      const xFrameOptions = (
-        (response.headers['x-frame-options'] as string) || ''
-      ).toLowerCase();
-      const csp = (
-        (response.headers['content-security-policy'] as string) || ''
-      ).toLowerCase();
-
-      let isEmbeddable = true;
-      let reason = '';
-
-      if (xFrameOptions === 'deny' || xFrameOptions === 'sameorigin') {
-        isEmbeddable = false;
-        reason = `Site specifies 'X-Frame-Options: ${xFrameOptions.toUpperCase()}'.`;
-      } else if (csp.includes('frame-ancestors')) {
-        // Very basic check - if frame-ancestors is present and doesn't explicitly allow all or the current origin
-        // In a real scenario, we'd need to parse the CSP properly, but 'self' or 'none' are the most common blocks.
-        if (csp.includes("'self'") || csp.includes("'none'")) {
-          isEmbeddable = false;
-          reason =
-            'Site has a strict Content Security Policy (frame-ancestors).';
-        }
-      }
-
-      return {
-        isEmbeddable,
-        reason,
-        headers: {
-          'x-frame-options': xFrameOptions,
-          'content-security-policy': csp,
-        },
-      };
-    } catch (error: unknown) {
-      console.error('Compatibility Check Error:', error);
-      // Some sites block HEAD requests or have other issues
-      return {
-        isEmbeddable: true, // Assume okay if we can't check, but we'll flag the error
-        error: error instanceof Error ? error.message : 'Failed to check site',
-        uncertain: true,
-      };
-    }
-  }
-);
-
-// ---------------------------------------------------------------------------
-// Video Activity: Caption-based AI question generation
-// ---------------------------------------------------------------------------
-
-type VideoQuestionType = 'MC' | 'FIB' | 'MA';
-
-interface VideoActivityRequestData {
-  url: string;
-  /**
-   * Per-type question counts the teacher requested in the AI overlay. The
-   * server requests a slight surplus from Gemini, then post-validates and
-   * trims back to these quotas so hallucinations get dropped instead of
-   * surfacing to the user.
-   */
-  typeCounts?: Partial<Record<VideoQuestionType, number>>;
-  /**
-   * Legacy single-count field. Older deployed clients send this instead of
-   * `typeCounts`. Server treats it as `{ MC: questionCount }` during a
-   * staggered functions/frontend deploy so in-flight calls don't break.
-   */
-  questionCount?: number;
-  /**
-   * Total video duration in seconds, captured client-side from the YouTube
-   * IFrame player. Used as a hard upper bound on emitted timestamps — any
-   * question Gemini returns with a timestamp beyond this is dropped.
-   * Optional because Gemini will still produce reasonable timestamps without
-   * it, but accuracy is much higher when provided.
-   */
-  durationSeconds?: number;
-}
-
-interface GeneratedVideoQuestion {
-  text: string;
-  timestamp: number;
-  type: VideoQuestionType;
-  correctAnswer: string;
-  incorrectAnswers: string[];
-  /** FIB-only: optional alternate accepted answer forms (e.g. ["color", "colour"]). */
-  acceptableVariants?: string[];
-  timeLimit: number;
-}
-
-interface GeneratedVideoActivity {
-  title: string;
-  questions: GeneratedVideoQuestion[];
-  /**
-   * Optional: `true` when the Cloud Function couldn't read the admin model
-   * config from Firestore and fell back to hardcoded defaults. Lets the
-   * client surface a one-time admin notice without polluting every call.
-   */
-  _modelConfigUsedFallback?: boolean;
-}
-
-// ---------------------------------------------------------------------------
-// Video Activity / Quiz AI helpers
-//
-// These helpers are exported (named) but not re-exported through the Firebase
-// callables list — they're internal to this file's two AI generators. Keeping
-// them as pure functions makes the prompt + post-validation logic testable in
-// isolation (no Firestore, no Gemini calls).
-// ---------------------------------------------------------------------------
-
-/** Order matters: this is the per-type emit order in prompts. */
-const VIDEO_QUESTION_TYPES: VideoQuestionType[] = ['MC', 'FIB', 'MA'];
-const QUIZ_QUESTION_TYPES: QuizGenType[] = [
-  'MC',
-  'FIB',
-  'Matching',
-  'Ordering',
-];
-
-const VIDEO_TYPE_LABEL: Record<VideoQuestionType, string> = {
-  MC: 'Multiple Choice',
-  FIB: 'Fill in the Blank',
-  MA: 'Multi-Answer (select multiple)',
-};
-
-const QUIZ_TYPE_LABEL: Record<QuizGenType, string> = {
-  MC: 'Multiple Choice',
-  FIB: 'Fill in the Blank',
-  Matching: 'Matching pairs',
-  Ordering: 'Sequence ordering',
-};
-
-interface NormalizedCounts<T extends string> {
-  counts: Record<T, number>;
-  total: number;
-}
-
-/** Clamp negative/non-finite counts to zero and compute a total. */
-function normalizeTypeCounts<T extends string>(
-  allowed: readonly T[],
-  raw: Partial<Record<T, number>> | undefined
-): NormalizedCounts<T> {
-  const counts = {} as Record<T, number>;
-  let total = 0;
-  for (const t of allowed) {
-    const n = Number(raw?.[t]);
-    const clean = Number.isFinite(n) && n > 0 ? Math.floor(n) : 0;
-    counts[t] = clean;
-    total += clean;
-  }
-  return { counts, total };
-}
-
-/**
- * Ask Gemini for a surplus so hallucinations can be filtered without
- * leaving the teacher short on questions. ~30% extra plus a minimum of +2,
- * capped at 60 so a pathological per-type request (15 of each of 4 quiz
- * types → 60 requested → 80 with surplus) doesn't balloon the prompt past
- * what Gemini can usefully reason about. The per-type stepper UI caps each
- * type at 15, so 60 raw is the worst case we should encounter in practice.
- */
-function bufferedRequestCount(requested: number): number {
-  if (requested <= 0) return 0;
-  return Math.min(60, Math.ceil(requested * 1.3) + 2);
-}
-
-/**
- * Back-compat shim for in-flight video-activity callers that haven't
- * received the new bundle yet. If the request only contains the legacy
- * `questionCount` field (or both fields are empty), treat it as
- * `{ MC: questionCount }` so we don't reject a payload from an older
- * frontend during a staggered Functions/frontend deploy.
- */
-function legacyVideoTypeCountsFallback(
-  typeCounts: Partial<Record<VideoQuestionType, number>> | undefined,
-  questionCount: number | undefined
-): Partial<Record<VideoQuestionType, number>> | undefined {
-  const supplied = typeCounts ?? {};
-  const hasAny = VIDEO_QUESTION_TYPES.some((t) => Number(supplied[t]) > 0);
-  if (hasAny) return supplied;
-  const legacy = Number(questionCount);
-  if (Number.isFinite(legacy) && legacy > 0) {
-    return { MC: Math.floor(legacy) };
-  }
-  return supplied;
-}
-
-function buildVideoActivityPrompt(
-  counts: Record<VideoQuestionType, number>,
-  total: number,
-  bufferedTotal: number,
-  durationSeconds: number | undefined
-): string {
-  const mixLines = VIDEO_QUESTION_TYPES.filter((t) => counts[t] > 0)
-    .map((t) => `   - ${counts[t]} ${VIDEO_TYPE_LABEL[t]} ("${t}")`)
-    .join('\n');
-
-  const durationLine =
-    durationSeconds && durationSeconds > 0
-      ? `The video is exactly ${Math.floor(durationSeconds)} seconds long. Every "timestamp" you emit MUST be an integer between 0 and ${Math.floor(durationSeconds)} inclusive. NEVER invent timestamps beyond ${Math.floor(durationSeconds)}.`
-      : 'Every "timestamp" you emit MUST be an integer second that falls within the actual video. Do not invent timestamps past the end of the video.';
-
-  return `You are an expert teacher creating a video comprehension activity.
-Watch the provided YouTube video and generate up to ${bufferedTotal} questions that check understanding of key concepts.
-
-TARGET QUESTION MIX (the teacher requested ${total} questions total):
-${mixLines}
-
-CRITICAL RULES:
-1. ${durationLine}
-2. Each question's "timestamp" should sit AT or JUST AFTER the moment the answer is discussed, so students hear the explanation before being prompted.
-3. Questions must be in ascending "timestamp" order.
-4. Every question must be grounded in something actually said or shown in the video. Do not paraphrase generic facts that aren't covered.
-5. Per-type shape (the "type" field decides which fields are required):
-   - "MC": "correctAnswer" is the one correct option (string). "incorrectAnswers" MUST contain exactly 3 plausible-but-wrong distractor strings. Leave "acceptableVariants" empty.
-   - "FIB": "correctAnswer" is the single canonical accepted answer (a short word or phrase students would type). "incorrectAnswers" MUST be an empty array. Optionally fill "acceptableVariants" with 0–3 additional accepted spellings/synonyms (e.g. ["color", "colour"]).
-   - "MA": "correctAnswer" is a pipe-separated list of the correct selections, e.g. "option1|option2". "incorrectAnswers" contains 2–4 distractor options shown alongside. Leave "acceptableVariants" empty.
-6. "timeLimit" should be 20–45 seconds (integer).
-7. Aim for the requested type counts above. If you cannot find enough good content for a type, emit fewer of that type rather than padding with another type or fabricating content.
-8. Return ONLY valid JSON matching the provided schema. No markdown fences, no commentary.`;
-}
-
-function buildQuizPrompt(
-  counts: Record<QuizGenType, number>,
-  total: number,
-  bufferedTotal: number,
-  userInput: string
-): string {
-  const mixLines = QUIZ_QUESTION_TYPES.filter((t) => counts[t] > 0)
-    .map((t) => `   - ${counts[t]} ${QUIZ_TYPE_LABEL[t]} ("${t}")`)
-    .join('\n');
-
-  return `You are an expert teacher creating a classroom quiz.
-Generate a quiz based on the topic or content provided within <topic> tags below. Produce up to ${bufferedTotal} questions to satisfy the requested mix (target ${total} total).
-
-TARGET QUESTION MIX:
-${mixLines}
-
-Per-type shape (the "type" field decides which fields are required):
-- "MC": "correctAnswer" is the one correct option. "incorrectAnswers" MUST contain exactly 3 plausible-but-wrong distractor strings.
-- "FIB": "correctAnswer" is the single canonical accepted answer (short word or phrase). "incorrectAnswers" MUST be an empty array.
-- "Matching": "correctAnswer" is a pipe-separated list of "term:definition" pairs, e.g. "Mars:fourth planet|Venus:second planet|Earth:third planet". Provide 3–6 pairs. "incorrectAnswers" MUST be an empty array.
-- "Ordering": "correctAnswer" is a pipe-separated list of items in their correct sequence, e.g. "First|Second|Third|Fourth". Provide 3–6 items. "incorrectAnswers" MUST be an empty array.
-
-Other rules:
-1. "timeLimit" should be 20–60 seconds depending on complexity (integer).
-2. Questions should progress from easier to harder.
-3. Aim for the requested type counts above. If you cannot find enough good content for a type, emit fewer of that type rather than padding with another type or fabricating content.
-4. Return ONLY valid JSON matching the provided schema. No markdown fences, no commentary.
-
-Topic/Content: <topic>${userInput}</topic>`;
-}
-
-function buildVideoActivityResponseSchema(): Schema {
-  return {
-    type: Type.OBJECT,
-    required: ['title', 'questions'],
-    properties: {
-      title: { type: Type.STRING },
-      questions: {
-        type: Type.ARRAY,
-        items: {
-          type: Type.OBJECT,
-          required: [
-            'text',
-            'timestamp',
-            'type',
-            'correctAnswer',
-            'incorrectAnswers',
-            'timeLimit',
-          ],
-          properties: {
-            text: { type: Type.STRING },
-            timestamp: { type: Type.INTEGER },
-            type: {
-              type: Type.STRING,
-              enum: ['MC', 'FIB', 'MA'],
-            },
-            correctAnswer: { type: Type.STRING },
-            incorrectAnswers: {
-              type: Type.ARRAY,
-              items: { type: Type.STRING },
-            },
-            acceptableVariants: {
-              type: Type.ARRAY,
-              items: { type: Type.STRING },
-            },
-            timeLimit: { type: Type.INTEGER },
-          },
-        },
-      },
-    },
-  };
-}
-
-function buildQuizResponseSchema(): Schema {
-  return {
-    type: Type.OBJECT,
-    required: ['title', 'questions'],
-    properties: {
-      title: { type: Type.STRING },
-      questions: {
-        type: Type.ARRAY,
-        items: {
-          type: Type.OBJECT,
-          required: [
-            'text',
-            'type',
-            'correctAnswer',
-            'incorrectAnswers',
-            'timeLimit',
-          ],
-          properties: {
-            text: { type: Type.STRING },
-            type: {
-              type: Type.STRING,
-              enum: ['MC', 'FIB', 'Matching', 'Ordering'],
-            },
-            correctAnswer: { type: Type.STRING },
-            incorrectAnswers: {
-              type: Type.ARRAY,
-              items: { type: Type.STRING },
-            },
-            timeLimit: { type: Type.INTEGER },
-          },
-        },
-      },
-    },
-  };
-}
-
-/**
- * Drop malformed entries, drop entries with timestamps outside the video,
- * bucket by type, and trim each bucket to its requested quota. Returns the
- * accepted questions sorted by timestamp.
- *
- * Exported (via export keyword) primarily so a future test file can exercise
- * the validator without needing a Firebase/Gemini harness.
- */
-export function validateAndBucketVideoQuestions(
-  raw: unknown,
-  counts: Record<VideoQuestionType, number>,
-  durationSeconds: number | undefined
-): GeneratedVideoQuestion[] {
-  if (!Array.isArray(raw)) return [];
-  const maxTs =
-    durationSeconds && durationSeconds > 0
-      ? Math.floor(durationSeconds)
-      : Infinity;
-
-  const buckets: Record<VideoQuestionType, GeneratedVideoQuestion[]> = {
-    MC: [],
-    FIB: [],
-    MA: [],
-  };
-
-  for (const item of raw) {
-    if (!item || typeof item !== 'object') continue;
-    const q = item as Record<string, unknown>;
-    const type = q.type;
-    if (type !== 'MC' && type !== 'FIB' && type !== 'MA') {
-      continue;
-    }
-    const text = typeof q.text === 'string' ? q.text.trim() : '';
-    if (!text) continue;
-    const timestamp =
-      typeof q.timestamp === 'number' ? Math.floor(q.timestamp) : NaN;
-    if (!Number.isFinite(timestamp) || timestamp < 0 || timestamp > maxTs) {
-      continue;
-    }
-    const correctAnswer =
-      typeof q.correctAnswer === 'string' ? q.correctAnswer.trim() : '';
-    if (!correctAnswer) continue;
-    const incorrectAnswers = Array.isArray(q.incorrectAnswers)
-      ? q.incorrectAnswers.filter(
-          (a): a is string => typeof a === 'string' && a.trim().length > 0
-        )
-      : [];
-    // The prompt asks for exactly 3 distractors for MC. Validator floor is
-    // 3 to match — with the `bufferedRequestCount` surplus we'd rather drop
-    // a thin question than render a watered-down 3-option MC.
-    if (type === 'MC' && incorrectAnswers.length < 3) continue;
-    // MA stores correct selections pipe-joined in `correctAnswer` and
-    // shows `incorrectAnswers` as the distractor options. A "MA" with no
-    // distractors degenerates into "pick the only choice"; the prompt
-    // asks for 2–4 distractors, so we require at least 2 of each side.
-    if (type === 'MA' && !isValidOrderingList(correctAnswer, 2)) continue;
-    if (type === 'MA' && incorrectAnswers.length < 2) continue;
-    const timeLimitRaw =
-      typeof q.timeLimit === 'number' ? Math.floor(q.timeLimit) : 30;
-    const timeLimit =
-      Number.isFinite(timeLimitRaw) && timeLimitRaw > 0 ? timeLimitRaw : 30;
-    const acceptableVariants =
-      type === 'FIB' && Array.isArray(q.acceptableVariants)
-        ? q.acceptableVariants.filter(
-            (a): a is string => typeof a === 'string' && a.trim().length > 0
-          )
-        : undefined;
-
-    const validated: GeneratedVideoQuestion = {
-      text,
-      timestamp,
-      type,
-      correctAnswer,
-      incorrectAnswers: type === 'FIB' ? [] : incorrectAnswers,
-      timeLimit,
-      ...(acceptableVariants && acceptableVariants.length > 0
-        ? { acceptableVariants }
-        : {}),
-    };
-
-    if (buckets[type].length < counts[type]) {
-      buckets[type].push(validated);
-    }
-  }
-
-  return [...buckets.MC, ...buckets.FIB, ...buckets.MA].sort(
-    (a, b) => a.timestamp - b.timestamp
-  );
-}
-
-/** Pipe-delimited Ordering payloads must have at least 2 non-empty entries. */
-function isValidOrderingList(value: string, minEntries: number): boolean {
-  const parts = value.split('|').filter((p) => p.trim().length > 0);
-  return parts.length >= minEntries;
-}
-
-/**
- * Matching payloads use `"term1:def1|term2:def2"` per the storage rules in
- * types.ts. The validator must reject pipe-only output (an Ordering-shaped
- * payload that Gemini mislabeled as Matching), so each pair MUST contain a
- * non-empty term and a non-empty definition separated by `:`.
- */
-function isValidMatchingList(value: string, minPairs: number): boolean {
-  const parts = value.split('|').filter((p) => p.trim().length > 0);
-  if (parts.length < minPairs) return false;
-  return parts.every((pair) => {
-    const colonIdx = pair.indexOf(':');
-    if (colonIdx <= 0) return false;
-    const term = pair.slice(0, colonIdx).trim();
-    const def = pair.slice(colonIdx + 1).trim();
-    return term.length > 0 && def.length > 0;
-  });
-}
-
-interface ValidatedQuizQuestion {
-  text: string;
-  type: QuizGenType;
-  correctAnswer: string;
-  incorrectAnswers: string[];
-  timeLimit: number;
-}
-
-/**
- * Quiz counterpart to `validateAndBucketVideoQuestions`. Unlike the video
- * flow (which sorts by `timestamp`), the quiz prompt asks for "easier to
- * harder" progression, so we preserve Gemini's original emit order while
- * still enforcing per-type quotas: iterate raw in order, accept each valid
- * question until that type's bucket fills, drop the rest.
- */
-export function validateAndBucketQuizQuestions(
-  raw: unknown,
-  counts: Record<QuizGenType, number>
-): ValidatedQuizQuestion[] {
-  if (!Array.isArray(raw)) return [];
-  const accepted: ValidatedQuizQuestion[] = [];
-  const filled: Record<QuizGenType, number> = {
-    MC: 0,
-    FIB: 0,
-    Matching: 0,
-    Ordering: 0,
-  };
-
-  for (const item of raw) {
-    if (!item || typeof item !== 'object') continue;
-    const q = item as Record<string, unknown>;
-    const type = q.type;
-    if (
-      type !== 'MC' &&
-      type !== 'FIB' &&
-      type !== 'Matching' &&
-      type !== 'Ordering'
-    ) {
-      continue;
-    }
-    if (filled[type] >= counts[type]) continue;
-    const text = typeof q.text === 'string' ? q.text.trim() : '';
-    if (!text) continue;
-    const correctAnswer =
-      typeof q.correctAnswer === 'string' ? q.correctAnswer.trim() : '';
-    if (!correctAnswer) continue;
-    const incorrectAnswers = Array.isArray(q.incorrectAnswers)
-      ? q.incorrectAnswers.filter(
-          (a): a is string => typeof a === 'string' && a.trim().length > 0
-        )
-      : [];
-    // The prompts ask for exactly 3 distractors (MC) and 3+ pairs/items
-    // (Matching, Ordering). Validator floors match — with the
-    // `bufferedRequestCount` surplus we'd rather drop a thin question than
-    // render a 3-option MC or a 2-pair Matching that the teacher didn't
-    // ask for.
-    if (type === 'MC' && incorrectAnswers.length < 3) continue;
-    if (type === 'Matching' && !isValidMatchingList(correctAnswer, 3)) {
-      continue;
-    }
-    if (type === 'Ordering' && !isValidOrderingList(correctAnswer, 3)) {
-      continue;
-    }
-    const timeLimitRaw =
-      typeof q.timeLimit === 'number' ? Math.floor(q.timeLimit) : 30;
-    const timeLimit =
-      Number.isFinite(timeLimitRaw) && timeLimitRaw > 0 ? timeLimitRaw : 30;
-
-    accepted.push({
-      text,
-      type,
-      correctAnswer,
-      incorrectAnswers: type === 'MC' ? incorrectAnswers : [],
-      timeLimit,
-    });
-    filled[type] += 1;
-  }
-
-  return accepted;
-}
-
-/**
- * Uses Gemini's multimodal video understanding to analyze a YouTube video
- * and generate timestamped multiple-choice questions.
- */
-export const generateVideoActivity = onCall(
-  {
-    memory: '1GiB',
-    timeoutSeconds: 300,
-    secrets: [GEMINI_API_KEY],
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request): Promise<GeneratedVideoActivity> => {
-    const data = request.data as VideoActivityRequestData;
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    const uid = request.auth.uid;
-    const email = request.auth.token.email;
-
-    if (!email) {
-      throw new HttpsError(
-        'invalid-argument',
-        'User must have an email associated with their account.'
-      );
-    }
-
-    const { url, typeCounts, questionCount, durationSeconds } = data;
-
-    if (!url || typeof url !== 'string') {
-      throw new HttpsError(
-        'invalid-argument',
-        'A valid YouTube URL is required.'
-      );
-    }
-
-    const effectiveTypeCounts = legacyVideoTypeCountsFallback(
-      typeCounts,
-      questionCount
-    );
-    const { counts, total } = normalizeTypeCounts<VideoQuestionType>(
-      VIDEO_QUESTION_TYPES,
-      effectiveTypeCounts
-    );
-    if (total <= 0) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Select at least one question to generate.'
-      );
-    }
-    const bufferedTotal = bufferedRequestCount(total);
-
-    // Extract video ID from URL
-    const videoIdMatch = url.match(
-      /(?:youtu\.be\/|youtube\.com\/(?:embed\/|v\/|watch\?v=|watch\?.+&v=))([^&]{11})/
-    );
-    const videoId = videoIdMatch?.[1];
-
-    if (!videoId) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Could not extract a video ID from the provided URL. Please paste a valid YouTube link.'
-      );
-    }
-
-    const cleanDuration =
-      typeof durationSeconds === 'number' &&
-      Number.isFinite(durationSeconds) &&
-      durationSeconds > 0
-        ? Math.floor(durationSeconds)
-        : undefined;
-
-    const db = admin.firestore();
-
-    // Check if user is an admin (unlimited)
-    const adminDoc = await db
-      .collection('admins')
-      .doc(email.toLowerCase())
-      .get();
-    const isAdmin = adminDoc.exists;
-
-    if (!isAdmin) {
-      // --- Check Overall Gemini Limit ---
-      const today = new Date().toISOString().split('T')[0];
-      const overallUsageRef = db.collection('ai_usage').doc(`${uid}_${today}`);
-
-      try {
-        await db.runTransaction(async (transaction) => {
-          const globalPermDoc = await transaction.get(
-            db.collection('global_permissions').doc('gemini-functions')
-          );
-          const globalPerm = globalPermDoc.data() as
-            | GlobalPermission
-            | undefined;
-
-          if (globalPerm && !globalPerm.enabled) {
-            throw new HttpsError(
-              'permission-denied',
-              'Gemini functions are currently disabled by an administrator.'
-            );
-          }
-
-          const overallLimitEnabled =
-            globalPerm?.config?.dailyLimitEnabled !== false;
-          const overallLimit = globalPerm?.config?.dailyLimit ?? 20;
-
-          const overallUsageDoc = await transaction.get(overallUsageRef);
-          const currentOverallUsage =
-            (overallUsageDoc.data()?.count as number) || 0;
-
-          if (overallLimitEnabled && currentOverallUsage >= overallLimit) {
-            throw new HttpsError(
-              'resource-exhausted',
-              `Daily AI usage limit reached (${overallLimit} generations). Please try again tomorrow.`
-            );
-          }
-
-          transaction.set(
-            overallUsageRef,
-            {
-              count: currentOverallUsage + 1,
-              email,
-              lastUsed: admin.firestore.FieldValue.serverTimestamp(),
-            },
-            { merge: true }
-          );
-        });
-      } catch (error) {
-        if (error instanceof HttpsError) throw error;
-        console.error('Usage check error:', error);
-        throw new HttpsError('internal', 'Failed to verify AI usage limits.');
-      }
-    }
-
-    const apiKey = GEMINI_API_KEY.value();
-    if (!apiKey) {
-      throw new HttpsError(
-        'internal',
-        'Gemini API Key is missing on the server.'
-      );
-    }
-
-    // Read model config from Firestore
-    const geminiConfig = await getGeminiModelConfig(db);
-    const videoModel = geminiConfig.standardModel;
-
-    const ai = new GoogleGenAI({ apiKey });
-
-    const systemPrompt = buildVideoActivityPrompt(
-      counts,
-      total,
-      bufferedTotal,
-      cleanDuration
-    );
-
-    try {
-      const result = await ai.models.generateContent({
-        model: videoModel,
-        contents: [
-          {
-            role: 'user',
-            parts: [
-              { text: systemPrompt },
-              {
-                fileData: {
-                  fileUri: `https://www.youtube.com/watch?v=${videoId}`,
-                  mimeType: 'video/mp4',
-                },
-              },
-            ],
-          },
-        ],
-        config: {
-          responseMimeType: 'application/json',
-          responseSchema: buildVideoActivityResponseSchema(),
-        },
-      });
-
-      const text = result.text;
-      if (!text) throw new Error('Empty response from AI');
-
-      const parsed = parseGeminiJson<{
-        title?: unknown;
-        questions?: unknown;
-      }>(text);
-
-      if (!parsed.title || typeof parsed.title !== 'string') {
-        throw new Error('Invalid response structure from AI (missing title).');
-      }
-
-      const validatedQuestions = validateAndBucketVideoQuestions(
-        parsed.questions,
-        counts,
-        cleanDuration
-      );
-
-      if (validatedQuestions.length === 0) {
-        throw new Error(
-          'The AI did not produce any usable questions for this video. Try a different video or adjust the question mix.'
-        );
-      }
-
-      return {
-        title: parsed.title,
-        questions: validatedQuestions,
-        _modelConfigUsedFallback: geminiConfig.usedFallback,
-      };
-    } catch (error: unknown) {
-      console.error('[generateVideoActivity] Gemini error:', error);
-      if (error instanceof HttpsError) throw error;
-      const detail = error instanceof Error ? error.message : 'unknown error';
-      throw new HttpsError(
-        'internal',
-        `AI generation failed (model: ${videoModel}): ${detail}`
-      );
-    }
-  }
-);
-
-// ---------------------------------------------------------------------------
-// Video Activity: Admin-gated Gemini audio transcription fallback
-// For videos that do not have captions available.
-// ---------------------------------------------------------------------------
-
-interface AudioTranscriptionRequestData {
-  url: string;
-  /** See `VideoActivityRequestData.typeCounts`. */
-  typeCounts?: Partial<Record<VideoQuestionType, number>>;
-  /** See `VideoActivityRequestData.questionCount`. */
-  questionCount?: number;
-  /** See `VideoActivityRequestData.durationSeconds`. */
-  durationSeconds?: number;
-}
-
-interface AudioTranscriptionPermConfig {
-  dailyLimit?: number;
-  dailyLimitEnabled?: boolean;
-  model?: string;
-}
-
-interface AudioTranscriptionPerm {
-  enabled: boolean;
-  accessLevel: 'admin' | 'beta' | 'all';
-  betaUsers?: string[];
-  config?: AudioTranscriptionPermConfig;
-}
-
-/**
- * Admin-only fallback: uses Gemini multimodal to transcribe a video that has
- * no captions, then generates timestamped quiz questions.
- *
- * Gated behind global_permissions/video-activity-audio-transcription.
- * Separate daily usage counter to control costs independently.
- *
- * NOTE: Audio/video extraction from YouTube may be subject to YouTube Terms of
- * Service. This feature is disabled by default and must be explicitly enabled
- * by an administrator who accepts the associated risk.
- */
-export const transcribeVideoWithGemini = onCall(
-  {
-    memory: '1GiB',
-    timeoutSeconds: 300,
-    secrets: [GEMINI_API_KEY],
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request): Promise<GeneratedVideoActivity> => {
-    const data = request.data as AudioTranscriptionRequestData;
-    if (!request.auth) {
-      throw new HttpsError(
-        'unauthenticated',
-        'The function must be called while authenticated.'
-      );
-    }
-
-    const uid = request.auth.uid;
-    const email = request.auth.token.email;
-
-    if (!email) {
-      throw new HttpsError(
-        'invalid-argument',
-        'User must have an email associated with their account.'
-      );
-    }
-
-    const db = admin.firestore();
-
-    // Check feature permission — admin-gated, off by default
-    const permDoc = await db
-      .collection('global_permissions')
-      .doc('video-activity-audio-transcription')
-      .get();
-
-    if (!permDoc.exists) {
-      throw new HttpsError(
-        'permission-denied',
-        'Gemini audio transcription is not enabled. An administrator must enable it in Feature Permissions.'
-      );
-    }
-
-    const perm = permDoc.data() as AudioTranscriptionPerm;
-
-    if (!perm.enabled) {
-      throw new HttpsError(
-        'permission-denied',
-        'Gemini audio transcription is currently disabled.'
-      );
-    }
-
-    // Check admin status
-    const adminDoc = await db
-      .collection('admins')
-      .doc(email.toLowerCase())
-      .get();
-    const isAdmin = adminDoc.exists;
-
-    if (!isAdmin) {
-      if (perm.accessLevel === 'admin') {
-        throw new HttpsError(
-          'permission-denied',
-          'Gemini audio transcription is restricted to administrators.'
-        );
-      }
-      if (
-        perm.accessLevel === 'beta' &&
-        !perm.betaUsers?.includes(email.toLowerCase())
-      ) {
-        throw new HttpsError(
-          'permission-denied',
-          'You do not have access to Gemini audio transcription.'
-        );
-      }
-    }
-
-    if (!isAdmin) {
-      // --- Dual Limit Check (Overall + Specific) ---
-      const today = new Date().toISOString().split('T')[0];
-      const overallUsageRef = db.collection('ai_usage').doc(`${uid}_${today}`);
-      const specificUsageRef = db
-        .collection('ai_usage')
-        .doc(`${uid}_video-activity-audio-transcription_${today}`);
-
-      try {
-        await db.runTransaction(async (transaction) => {
-          // 1. Check Overall Limit
-          const globalPermDoc = await transaction.get(
-            db.collection('global_permissions').doc('gemini-functions')
-          );
-          const globalPerm = globalPermDoc.data() as
-            | GlobalPermission
-            | undefined;
-          const overallLimitEnabled =
-            globalPerm?.config?.dailyLimitEnabled !== false;
-          const overallLimit = globalPerm?.config?.dailyLimit ?? 20;
-
-          const overallUsageDoc = await transaction.get(overallUsageRef);
-          const currentOverallUsage =
-            (overallUsageDoc.data()?.count as number) || 0;
-
-          if (overallLimitEnabled && currentOverallUsage >= overallLimit) {
-            throw new HttpsError(
-              'resource-exhausted',
-              `Daily AI usage limit reached (${overallLimit} generations). Please try again tomorrow.`
-            );
-          }
-
-          // 2. Check Specific Transcription Limit
-          const specLimitEnabled = perm.config?.dailyLimitEnabled !== false;
-          const specLimit = perm.config?.dailyLimit ?? 5;
-
-          const specUsageDoc = await transaction.get(specificUsageRef);
-          const currentSpecUsage = (specUsageDoc.data()?.count as number) || 0;
-
-          if (specLimitEnabled && currentSpecUsage >= specLimit) {
-            throw new HttpsError(
-              'resource-exhausted',
-              `Daily audio transcription limit reached (${specLimit} per day). Please try again tomorrow.`
-            );
-          }
-
-          // 3. Increment Both
-          transaction.set(
-            overallUsageRef,
-            {
-              count: currentOverallUsage + 1,
-              email,
-              lastUsed: admin.firestore.FieldValue.serverTimestamp(),
-            },
-            { merge: true }
-          );
-
-          transaction.set(
-            specificUsageRef,
-            {
-              count: currentSpecUsage + 1,
-              email,
-              lastUsed: admin.firestore.FieldValue.serverTimestamp(),
-            },
-            { merge: true }
-          );
-        });
-      } catch (error) {
-        if (error instanceof HttpsError) throw error;
-        console.error('Transcription usage check error:', error);
-        throw new HttpsError(
-          'internal',
-          'Failed to verify audio transcription usage limits.'
-        );
-      }
-    }
-
-    const { url, typeCounts, durationSeconds } = data;
-
-    const { counts, total } = normalizeTypeCounts<VideoQuestionType>(
-      VIDEO_QUESTION_TYPES,
-      typeCounts
-    );
-    if (total <= 0) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Select at least one question to generate.'
-      );
-    }
-    const bufferedTotal = bufferedRequestCount(total);
-
-    const videoIdMatch = url.match(
-      /(?:youtu\.be\/|youtube\.com\/(?:embed\/|v\/|watch\?v=|watch\?.+&v=))([^&]{11})/
-    );
-    const videoId = videoIdMatch?.[1];
-
-    if (!videoId) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Could not extract a video ID from the provided URL.'
-      );
-    }
-
-    const cleanDuration =
-      typeof durationSeconds === 'number' &&
-      Number.isFinite(durationSeconds) &&
-      durationSeconds > 0
-        ? Math.floor(durationSeconds)
-        : undefined;
-
-    const apiKey = GEMINI_API_KEY.value();
-    if (!apiKey) {
-      throw new HttpsError(
-        'internal',
-        'Gemini API Key is missing on the server.'
-      );
-    }
-
-    // Use the YouTube video URL directly with Gemini's video understanding
-    const model = perm.config?.model ?? 'gemini-3.1-flash-lite-preview';
-    const ai = new GoogleGenAI({ apiKey });
-
-    const systemPrompt = buildVideoActivityPrompt(
-      counts,
-      total,
-      bufferedTotal,
-      cleanDuration
-    );
-
-    try {
-      const result = await ai.models.generateContent({
-        model,
-        contents: [
-          {
-            role: 'user',
-            parts: [
-              { text: systemPrompt },
-              {
-                fileData: {
-                  fileUri: `https://www.youtube.com/watch?v=${videoId}`,
-                  mimeType: 'video/mp4',
-                },
-              },
-            ],
-          },
-        ],
-        config: {
-          responseMimeType: 'application/json',
-          responseSchema: buildVideoActivityResponseSchema(),
-        },
-      });
-
-      const text = result.text;
-      if (!text) throw new Error('Empty response from AI');
-
-      const parsed = parseGeminiJson<{
-        title?: unknown;
-        questions?: unknown;
-      }>(text);
-
-      if (!parsed.title || typeof parsed.title !== 'string') {
-        throw new Error('Invalid response structure from AI (missing title).');
-      }
-
-      const validatedQuestions = validateAndBucketVideoQuestions(
-        parsed.questions,
-        counts,
-        cleanDuration
-      );
-
-      if (validatedQuestions.length === 0) {
-        throw new Error(
-          'The AI did not produce any usable questions for this video. Try a different video or adjust the question mix.'
-        );
-      }
-
-      return {
-        title: parsed.title,
-        questions: validatedQuestions,
-      };
-    } catch (error: unknown) {
-      console.error('[transcribeVideoWithGemini] Gemini error:', error);
-      if (error instanceof HttpsError) throw error;
-      const detail = error instanceof Error ? error.message : 'unknown error';
-      const msg = `AI generation failed (model: ${model}): ${detail}`;
-      throw new HttpsError('internal', msg);
-    }
-  }
-);
-
-// ─── Guided Learning Generation (Admin Only) ─────────────────────────────────
-
-interface GuidedLearningStep {
-  id: string;
-  xPct: number;
-  yPct: number;
-  imageIndex?: number;
-  label?: string;
-  interactionType: string;
-  hideStepNumber?: boolean;
-  showOverlay?: 'none' | 'popover' | 'tooltip' | 'banner';
-  text?: string;
-  panZoomScale?: number;
-  spotlightRadius?: number;
-  question?: {
-    type: string;
-    text: string;
-    choices?: string[];
-    correctAnswer?: string;
-    matchingPairs?: { left: string; right: string }[];
-    sortingItems?: string[];
-  };
-  autoAdvanceDuration?: number;
-}
-
-interface GeneratedGuidedLearning {
-  suggestedTitle: string;
-  suggestedMode: string;
-  steps: GuidedLearningStep[];
-  /** See `GeneratedVideoActivity._modelConfigUsedFallback`. */
-  _modelConfigUsedFallback?: boolean;
-}
-
-interface GuidedLearningImageInput {
-  base64: string;
-  mimeType: string;
-  caption?: string;
-}
-
-export const generateGuidedLearning = onCall(
-  {
-    memory: '512MiB',
-    timeoutSeconds: 120,
-    secrets: [GEMINI_API_KEY],
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = request.data as {
-      images?: GuidedLearningImageInput[];
-      prompt?: string;
-      // Legacy single-image shape — accepted for backward compatibility.
-      imageBase64?: string;
-      mimeType?: string;
-    };
-    // Admin only
-    const uid = request.auth?.uid;
-    if (!uid) {
-      throw new HttpsError(
-        'unauthenticated',
-        'Must be authenticated to use this feature.'
-      );
-    }
-
-    const userEmail = request.auth?.token.email;
-    if (!userEmail) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Authenticated user must have an email address.'
-      );
-    }
-    const db = admin.firestore();
-    const adminDoc = await db
-      .collection('admins')
-      .doc(userEmail.toLowerCase())
-      .get();
-    if (!adminDoc.exists) {
-      throw new HttpsError(
-        'permission-denied',
-        'Admin access required to use AI generation.'
-      );
-    }
-
-    const { prompt } = data;
-    const images: GuidedLearningImageInput[] =
-      Array.isArray(data.images) && data.images.length > 0
-        ? data.images
-        : data.imageBase64 && data.mimeType
-          ? [{ base64: data.imageBase64, mimeType: data.mimeType }]
-          : [];
-
-    if (images.length === 0) {
-      throw new HttpsError(
-        'invalid-argument',
-        'At least one image is required.'
-      );
-    }
-
-    const MAX_IMAGES = 10;
-    const MAX_TOTAL_RAW_BYTES = 20 * 1024 * 1024; // 20 MB raw (~27 MB base64)
-    if (images.length > MAX_IMAGES) {
-      throw new HttpsError(
-        'invalid-argument',
-        `Too many images — please limit to ${MAX_IMAGES} per request.`
-      );
-    }
-
-    let totalRawBytes = 0;
-    for (const img of images) {
-      if (!img.base64 || !img.mimeType) {
-        throw new HttpsError(
-          'invalid-argument',
-          'Every image must include base64 data and mimeType.'
-        );
-      }
-      // Base64 decodes to ~0.75 bytes per char (minus padding).
-      totalRawBytes += Math.ceil((img.base64.length * 3) / 4);
-    }
-    if (totalRawBytes > MAX_TOTAL_RAW_BYTES) {
-      throw new HttpsError(
-        'invalid-argument',
-        'Image payload is too large. Please use fewer or smaller images (under 20 MB total).'
-      );
-    }
-
-    const apiKey = GEMINI_API_KEY.value();
-    if (!apiKey) {
-      throw new HttpsError('internal', 'AI service is not configured.');
-    }
-
-    // Read model config from Firestore
-    const geminiConfig = await getGeminiModelConfig(db);
-    const guidedLearningModel = geminiConfig.advancedModel;
-
-    try {
-      const ai = new GoogleGenAI({ apiKey });
-
-      const imageCount = images.length;
-      const maxIndex = imageCount - 1;
-      const systemInstruction = `You are an educational content creator helping teachers build interactive guided learning experiences.
-Analyze the provided image(s) and generate a guided learning experience as a JSON object.
-
-Return ONLY valid JSON with this exact structure:
-{
-  "suggestedTitle": "string",
-  "suggestedMode": "structured" | "guided" | "explore",
-  "steps": [
-    {
-      "id": "unique-string",
-      "xPct": number (0-100),
-      "yPct": number (0-100),
-      "label": "string",
-      "interactionType": "text-popover" | "tooltip" | "pan-zoom" | "spotlight" | "pan-zoom-spotlight" | "question",
-      "imageIndex": number (0-based index into the provided images, 0..${maxIndex}),
-      "hideStepNumber": boolean (optional),
-      "showOverlay": "none" | "popover" | "tooltip" | "banner" (for pan-zoom, spotlight, pan-zoom-spotlight),
-      "text": "string (for text-popover/tooltip)",
-      "panZoomScale": number (1.5-4, for pan-zoom and pan-zoom-spotlight),
-      "spotlightRadius": number (10-40, for spotlight and pan-zoom-spotlight),
-      "autoAdvanceDuration": number (seconds),
-      "question": {
-        "type": "multiple-choice" | "matching" | "sorting",
-        "text": "string",
-        "choices": ["string"] (MC: include correct + 3 incorrect),
-        "correctAnswer": "string (MC: must match one choice)",
-        "matchingPairs": [{"left": "string", "right": "string"}],
-        "sortingItems": ["string"] (in correct order)
-      }
-    }
-  ]
-}
-
-Guidelines:
-- You have been given ${imageCount} image${imageCount === 1 ? '' : 's'} (imageIndex ${imageCount === 1 ? '0' : `0..${maxIndex}`}).
-- Each step's imageIndex MUST be the 0-based position of the image it refers to.
-- Distribute steps across the images in a pedagogically meaningful order; do not cluster everything on image 0 unless only one image was provided.
-- Respect any per-image notes the teacher provided (sent as text before each image).
-- Create 4-8 meaningful steps per image that guide learners through the content (scale total step count with image count, cap at ~${Math.min(24, imageCount * 6)}).
-- Use text-popover for key concepts, spotlight to highlight areas, pan-zoom to zoom in on details, pan-zoom-spotlight when both are useful, questions to check understanding.
-- Place hotspots at meaningful locations on the image (xPct/yPct as percentages 0-100, relative to the image they reference).
-- Include at least 1 question step for comprehension checking.
-- Make content educational and age-appropriate.
-- Set autoAdvanceDuration to 5-15 seconds for non-question steps in guided mode.`;
-
-      const userPromptHeader = prompt
-        ? `Additional instructions: ${sanitizePrompt(prompt)}`
-        : 'Analyze the image(s) below and create an engaging guided learning experience.';
-
-      const parts: {
-        text?: string;
-        inlineData?: { mimeType: string; data: string };
-      }[] = [{ text: userPromptHeader }];
-
-      images.forEach((img, index) => {
-        const caption = img.caption ? sanitizePrompt(img.caption) : '';
-        const header = caption
-          ? `Image ${index} notes: ${caption}`
-          : `Image ${index}:`;
-        parts.push({ text: header });
-        parts.push({
-          inlineData: {
-            mimeType: img.mimeType,
-            data: img.base64,
-          },
-        });
-      });
-
-      const response = await ai.models.generateContent({
-        model: guidedLearningModel,
-        contents: [
-          {
-            role: 'user',
-            parts,
-          },
-        ],
-        config: {
-          systemInstruction,
-          responseMimeType: 'application/json',
-        },
-      });
-
-      const rawText = response.text ?? '';
-      const parsed = parseGeminiJson<GeneratedGuidedLearning>(rawText);
-
-      if (
-        !parsed.suggestedTitle ||
-        !Array.isArray(parsed.steps) ||
-        parsed.steps.length === 0
-      ) {
-        throw new Error('Invalid response structure from AI');
-      }
-
-      // Ensure all steps have IDs and imageIndex values fall within range.
-      parsed.steps = parsed.steps.map((step, i) => ({
-        ...step,
-        id: step.id || `step-${i + 1}-${Date.now()}`,
-        imageIndex:
-          typeof step.imageIndex === 'number' &&
-          step.imageIndex >= 0 &&
-          step.imageIndex <= maxIndex
-            ? step.imageIndex
-            : 0,
-      }));
-
-      return {
-        ...parsed,
-        _modelConfigUsedFallback: geminiConfig.usedFallback,
-      };
-    } catch (error: unknown) {
-      console.error('[generateGuidedLearning] Gemini error:', error);
-      const detail = error instanceof Error ? error.message : 'unknown error';
-      const msg = `AI generation failed (model: ${guidedLearningModel}): ${detail}`;
-      throw new HttpsError('internal', msg);
-    }
-  }
-);
-
-/**
- * Cloud Function to fetch administrative analytics.
- *
- * Read-only hot path: returns the snapshot at
- * `/organizations/{orgId}/analytics/snapshot` (written nightly by
- * `recomputeAdminAnalytics`). The previous implementation computed the
- * payload inline on every call, doing two unbounded reads
- * (`collectionGroup('dashboards').stream()` + `collection('ai_usage').stream()`)
- * that dominated the bill once user counts grew. The compute helper itself
- * still lives in `adminAnalyticsCompute.ts`; it now runs once a day from the
- * scheduled job in `adminAnalyticsSnapshot.ts` instead of on every page load.
- *
- * Memory is held at 1 GiB rather than the legacy 4 GiB because no streaming
- * reads happen here anymore — only the snapshot doc read plus the auth/authz
- * doc reads. Could probably drop further; leaving headroom for the JSON
- * serialization of large snapshots.
- */
-export const adminAnalytics = onRequest(
-  {
-    memory: '1GiB',
-    timeoutSeconds: 30,
-    cors: ALLOWED_ORIGINS,
-    invoker: 'public',
-  },
-  async (req, res) => {
-    // Correlation id for log triage. Emitted on the response (body +
-    // X-Request-Id header) and threaded through every `[getAdminAnalytics]`
-    // log line so a Cloud Logging alert can be pivoted back to the exact
-    // client-visible response.
-    const requestId = randomUUID();
-    res.setHeader('X-Request-Id', requestId);
-
-    // 1. Verify caller is authenticated via Bearer token
-    const authHeader = req.headers.authorization;
-    if (!authHeader?.startsWith('Bearer ')) {
-      console.error('[getAdminAnalytics] Unauthenticated access attempt', {
-        requestId,
-      });
-      res.status(401).json({ error: 'unauthenticated', requestId });
-      return;
-    }
-
-    let email: string;
-    try {
-      const idToken = authHeader.split('Bearer ')[1];
-      const decodedToken = await admin.auth().verifyIdToken(idToken);
-      if (!decodedToken.email) {
-        res.status(401).json({ error: 'unauthenticated', requestId });
-        return;
-      }
-      email = decodedToken.email.toLowerCase();
-    } catch {
-      res.status(401).json({ error: 'unauthenticated', requestId });
-      return;
-    }
-
-    // 1b. Require an orgId in the request body so analytics can be scoped to
-    // a single tenant. The previous behavior listed every Firebase Auth user
-    // globally, which leaked foreign-domain accounts into the calling admin's
-    // analytics view.
-    const rawBody = req.body as { orgId?: unknown } | undefined;
-    const orgId =
-      rawBody && typeof rawBody.orgId === 'string' ? rawBody.orgId.trim() : '';
-    if (!orgId) {
-      res.status(400).json({
-        error: 'invalid-argument',
-        message: 'orgId is required',
-        requestId,
-      });
-      return;
-    }
-
-    const db = admin.firestore();
-
-    // 2. Verify caller is authorized for the requested org. Two paths:
-    //   - Super admin: exists in `/admins/{email}`. May view any org.
-    //   - Org admin: has a member doc at `/organizations/{orgId}/members/{email}`
-    //     whose `roleId` is in the admin-tier set. Mirrors the role gating in
-    //     `assertCallerIsOrgAdmin` (organizationInvites.ts) but also admits
-    //     building_admin, since reading analytics is a lesser privilege than
-    //     inviting members.
-    const ORG_ADMIN_ROLE_IDS = new Set([
-      'super_admin',
-      'domain_admin',
-      'building_admin',
-    ]);
-    const [adminDoc, memberDoc] = await Promise.all([
-      db.collection('admins').doc(email).get(),
-      db.doc(`organizations/${orgId}/members/${email}`).get(),
-    ]);
-    const memberData = memberDoc.exists
-      ? (memberDoc.data() as { roleId?: unknown })
-      : undefined;
-    const memberRoleId =
-      typeof memberData?.roleId === 'string' ? memberData.roleId.trim() : '';
-    const isSuperAdmin = adminDoc.exists;
-    const isOrgAdmin = memberDoc.exists && ORG_ADMIN_ROLE_IDS.has(memberRoleId);
-    if (!isSuperAdmin && !isOrgAdmin) {
-      console.error('[getAdminAnalytics] Unauthorized access', {
-        requestId,
-        email,
-        orgId,
-      });
-      res.status(403).json({ error: 'permission-denied', requestId });
-      return;
-    }
-
-    try {
-      // Hot path is now snapshot-only. The compute logic that used to live
-      // here (~500 LOC) has moved to `adminAnalyticsCompute.ts` and runs
-      // once a day from the scheduled job in `adminAnalyticsSnapshot.ts`.
-      //
-      // Three reads total at this point: admin doc, member doc (both fetched
-      // above for the authz gate), and the snapshot doc. Compared to the
-      // pre-cache implementation this skips a `collectionGroup('dashboards')`
-      // stream over every dashboard in the database plus an `ai_usage`
-      // stream over every per-user counter — the two reads that were
-      // dominating the bill.
-      const snapshot = await readAnalyticsSnapshot(orgId);
-      if (!snapshot) {
-        // No snapshot yet — either a brand-new org pre-first-scheduled-run,
-        // or the snapshot was written under an older schemaVersion that the
-        // reader rejects. Either way, surface a 503 with a deterministic
-        // payload shape so the UI can show "Analytics will be ready after
-        // the next scheduled refresh" rather than spinning forever.
-        console.log('[getAdminAnalytics] no snapshot available', {
-          requestId,
-          orgId,
-        });
-        res.status(503).json({
-          error: 'not-yet-computed',
-          message:
-            'Analytics for this organization have not been computed yet. The next scheduled refresh runs at 5:00 AM Central daily.',
-          requestId,
-        });
-        return;
-      }
-
-      const { meta: payloadMeta, ...payloadWithoutMeta } = snapshot.payload;
-      res.json({
-        ...payloadWithoutMeta,
-        meta: {
-          ...(payloadMeta ?? {}),
-          computedAt: snapshot.computedAt,
-          nextRecomputeAt: snapshot.nextRecomputeAt,
-          computeDurationMs: snapshot.computeDurationMs,
-        },
-      });
-    } catch (err: unknown) {
-      console.error('[getAdminAnalytics] Error reading snapshot', {
-        requestId,
-        orgId,
-        error: err instanceof Error ? err.message : String(err),
-      });
-      const errorMessage =
-        err instanceof Error
-          ? err.message
-          : 'An internal error occurred reading analytics.';
-      res
-        .status(500)
-        .json({ error: 'internal', message: errorMessage, requestId });
-    }
-  }
-);
-
-// ---------------------------------------------------------------------------
-// Student identity (ClassLink-via-Google) — PII-free auth flow
-// ---------------------------------------------------------------------------
-//
-// Students launch SpartBoard from their ClassLink LaunchPad tile. Because
-// ClassLink is the district IdP and pushes identity into Google Workspace,
-// the student is already signed in with Google on the Chromebook. We use
-// Google Identity Services (GIS) client-side to obtain an ID token, verify
-// it server-side, then look up the student in ClassLink OneRoster and mint
-// a Firebase custom token whose UID is an HMAC pseudonym of the OneRoster
-// sourcedId. Email / name / sub / sourcedId are never persisted.
-//
-// Intentional design choices:
-//   - GIS + custom token (NOT signInWithPopup + GoogleAuthProvider) so that
-//     the Firebase Auth user record never receives email/displayName/photoURL.
-//   - Per-organization domain gating via existing
-//     /organizations/{orgId}/domains subcollection; a login with a domain
-//     not present (and verified) in any organization is rejected.
-//   - Per-assignment pseudonym = HMAC(SECRET, uid + assignmentId) so the
-//     server never needs the sourcedId after login. Teacher match-back
-//     recomputes it from the OneRoster roster at grading time.
-//   - NO PII logging. All catch blocks log class-of-failure only.
-
-interface OneRosterUserWithRole extends ClassLinkUser {
-  role?: string;
-  roles?: Array<{ role?: string; roleType?: string }>;
-}
-
-const STUDENT_LOGIN_CLASS_IDS_MAX = 20;
-
-function hmacSha256Hex(secret: string, message: string): string {
-  return CryptoJS.HmacSHA256(message, secret).toString(CryptoJS.enc.Hex);
-}
-
-function computeAssignmentPseudonym(
-  uid: string,
-  assignmentId: string,
-  hmacSecret: string
-): string {
-  return hmacSha256Hex(hmacSecret, `asn:${uid}:${assignmentId}`);
-}
-
-/**
- * Rejects emails that would break (or be injected into) an unquoted
- * OneRoster `filter=email='...'` string. Real Google-verified school emails
- * never contain `'` or `\`, so callers short-circuit to the standard
- * "not in roster" path rather than disclosing the guard's existence.
- */
-function isSafeEmailForOneRosterFilter(email: string): boolean {
-  return !/['\\]/.test(email);
-}
-
-function isOneRosterStudent(user: OneRosterUserWithRole): boolean {
-  if (user.role && user.role.toLowerCase() === 'student') return true;
-  if (Array.isArray(user.roles)) {
-    return user.roles.some((r) => {
-      const v = (r.role ?? r.roleType ?? '').toLowerCase();
-      return v === 'student' || v === 'primary';
-    });
-  }
-  return false;
-}
-
-/**
- * studentLoginV1
- *
- * Input:  { idToken: string }  — Google ID token from GIS on the client.
- * Output: { customToken, orgId, classCount } — client then calls
- *         signInWithCustomToken(customToken).
- *
- * Failure codes:
- *   - unauthenticated / invalid-argument: ID token missing or invalid.
- *   - permission-denied: email domain not registered with any organization.
- *   - not-found: student email not present in ClassLink OneRoster, or no
- *                classes enrolled, or account role is not 'student'.
- *   - internal: ClassLink API unreachable or server misconfigured.
- */
-export const studentLoginV1 = onCall(
-  {
-    memory: '256MiB',
-    cors: ALLOWED_ORIGINS,
-    secrets: [
-      CLASSLINK_CLIENT_ID,
-      CLASSLINK_CLIENT_SECRET,
-      CLASSLINK_TENANT_URL,
-      STUDENT_PSEUDONYM_HMAC_SECRET,
-      GOOGLE_OAUTH_CLIENT_ID,
-    ],
-    invoker: 'public',
-  },
-  async (request) => {
-    const rawIdToken = (request.data as { idToken?: unknown })?.idToken;
-    const idToken = typeof rawIdToken === 'string' ? rawIdToken : '';
-    if (!idToken) {
-      throw new HttpsError('invalid-argument', 'Missing idToken.');
-    }
-
-    const googleClientId = GOOGLE_OAUTH_CLIENT_ID.value();
-    const hmacSecret = STUDENT_PSEUDONYM_HMAC_SECRET.value();
-    const classlinkClientId = CLASSLINK_CLIENT_ID.value();
-    const classlinkClientSecret = CLASSLINK_CLIENT_SECRET.value();
-    const tenantUrl = CLASSLINK_TENANT_URL.value();
-    if (
-      !googleClientId ||
-      !hmacSecret ||
-      !classlinkClientId ||
-      !classlinkClientSecret ||
-      !tenantUrl
-    ) {
-      console.error(
-        '[studentLoginV1] Missing required server configuration (secrets).'
-      );
-      throw new HttpsError('internal', 'Server configuration missing.');
-    }
-
-    // 1. Verify the Google ID token signature and audience. The library also
-    //    validates `iss`, `exp`, and our expected `aud` in one call.
-    const oauthClient = new OAuth2Client();
-    let email: string;
-    let hd: string | undefined;
-    let emailVerified: boolean;
-    try {
-      const ticket = await oauthClient.verifyIdToken({
-        idToken,
-        audience: googleClientId,
-      });
-      const payload = ticket.getPayload();
-      if (!payload) {
-        throw new Error('no-payload');
-      }
-      email = typeof payload.email === 'string' ? payload.email : '';
-      hd = typeof payload.hd === 'string' ? payload.hd : undefined;
-      emailVerified = payload.email_verified === true;
-    } catch {
-      // Do not log token contents.
-      console.warn('[studentLoginV1] ID token verification failed.');
-      throw new HttpsError('unauthenticated', 'Invalid identity token.');
-    }
-    if (!email || !emailVerified) {
-      throw new HttpsError('unauthenticated', 'Email not verified by Google.');
-    }
-
-    // 2. Organization / domain gate. Prefer the `hd` claim (Workspace-issued),
-    //    but fall back to the email suffix since `hd` is not guaranteed on
-    //    every Workspace configuration.
-    const emailDomain = normalizeEmailDomain(email);
-    if (!emailDomain) {
-      throw new HttpsError('unauthenticated', 'Malformed email.');
-    }
-    const hdDomain = hd ? '@' + hd.toLowerCase() : null;
-
-    const db = admin.firestore();
-    let orgId = hdDomain ? await resolveOrgIdForDomain(db, hdDomain) : null;
-    if (!orgId) {
-      orgId = await resolveOrgIdForDomain(db, emailDomain);
-    }
-    if (!orgId) {
-      // Counter for monitoring alert on misconfiguration / unregistered schools.
-      console.warn('[studentLoginV1] students_rejected_domain');
-      throw new HttpsError(
-        'permission-denied',
-        'This SpartBoard is only available to schools that have signed up.'
-      );
-    }
-
-    // 2.5 Mock-class bypass. Admin-managed `testClasses` docs let us exercise
-    //     the end-to-end student SSO flow without provisioning the student in
-    //     ClassLink/OneRoster. If the email matches at least one testClasses
-    //     doc under this org, we short-circuit and mint a custom token whose
-    //     `classIds` are the testClasses doc ids — no OneRoster call is made.
-    //     Test uids are namespaced (`test:email`) so they never collide with
-    //     real OneRoster `sourcedId`-derived uids.
-    const emailLower = email.toLowerCase();
-    const testClassSnap = await db
-      .collection(`organizations/${orgId}/testClasses`)
-      .where('memberEmails', 'array-contains', emailLower)
-      .limit(STUDENT_LOGIN_CLASS_IDS_MAX)
-      .get();
-    if (!testClassSnap.empty) {
-      // Query is already bounded by `.limit(STUDENT_LOGIN_CLASS_IDS_MAX)` —
-      // no secondary slice needed.
-      const mockClassIds = testClassSnap.docs.map((d) => d.id);
-      // Monitoring counter — surface any prod use of the bypass.
-      console.warn('[studentLoginV1] test_bypass_used', { orgId });
-      const uid = computeStudentUid(`test:${emailLower}`, hmacSecret);
-      try {
-        const customToken = await admin.auth().createCustomToken(uid, {
-          studentRole: true,
-          orgId,
-          classIds: mockClassIds,
-        });
-        return { customToken, orgId, classCount: mockClassIds.length };
-      } catch (err) {
-        console.error(
-          '[studentLoginV1] createCustomToken failed (test bypass):',
-          err
-        );
-        throw new HttpsError('internal', 'Failed to mint auth token.');
-      }
-    }
-
-    // 3. ClassLink OneRoster lookup — fetch the student's sourcedId and
-    //    classes. Held in memory only, never written to Firestore.
-    const cleanTenantUrl = tenantUrl.replace(/\/$/, '');
-    let sourcedId: string;
-    let classIds: string[];
-    try {
-      if (!isSafeEmailForOneRosterFilter(email)) {
-        console.warn('[studentLoginV1] students_not_in_roster');
-        throw new HttpsError(
-          'not-found',
-          'No student record found in ClassLink roster.'
-        );
-      }
-      const usersBaseUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users`;
-      const userParams = { filter: `email='${email}'` };
-      const userHeaders = getOAuthHeaders(
-        usersBaseUrl,
-        userParams,
-        'GET',
-        classlinkClientId,
-        classlinkClientSecret
-      );
-      const userResp = await axios.get<{ users: OneRosterUserWithRole[] }>(
-        usersBaseUrl,
-        { params: userParams, headers: { ...userHeaders } }
-      );
-      const users = userResp.data.users ?? [];
-      const studentUser = users.find(isOneRosterStudent);
-      if (!studentUser) {
-        console.warn('[studentLoginV1] students_not_in_roster');
-        throw new HttpsError(
-          'not-found',
-          'No student record found in ClassLink roster.'
-        );
-      }
-      sourcedId = studentUser.sourcedId;
-
-      const classesUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users/${sourcedId}/classes`;
-      const classesHeaders = getOAuthHeaders(
-        classesUrl,
-        {},
-        'GET',
-        classlinkClientId,
-        classlinkClientSecret
-      );
-      const classesResp = await axios.get<{ classes: ClassLinkClass[] }>(
-        classesUrl,
-        { headers: { ...classesHeaders } }
-      );
-      classIds = (classesResp.data.classes ?? [])
-        .map((c) => c.sourcedId)
-        .filter((id): id is string => typeof id === 'string' && id.length > 0)
-        .slice(0, STUDENT_LOGIN_CLASS_IDS_MAX);
-    } catch (err) {
-      if (err instanceof HttpsError) throw err;
-      if (axios.isAxiosError(err)) {
-        console.error(
-          '[studentLoginV1] ClassLink request failed:',
-          err.response?.status
-        );
-      } else {
-        console.error('[studentLoginV1] ClassLink request failed.');
-      }
-      throw new HttpsError('internal', 'Roster service unavailable.');
-    }
-
-    // 4. Compute the stable opaque UID and mint the custom token with
-    //    the classIds claim that gates Firestore reads.
-    const uid = computeStudentUid(sourcedId, hmacSecret);
-
-    let customToken: string;
-    try {
-      customToken = await admin.auth().createCustomToken(uid, {
-        studentRole: true,
-        orgId,
-        classIds,
-      });
-    } catch (err) {
-      console.error('[studentLoginV1] createCustomToken failed:', err);
-      throw new HttpsError('internal', 'Failed to mint auth token.');
-    }
-
-    return { customToken, orgId, classCount: classIds.length };
-  }
-);
-
-/**
- * getAssignmentPseudonymV1
- *
- * Called by the authenticated student client when opening a specific
- * assignment. Returns the opaque pseudonym to write into the response doc:
- *
- *   pseudonym = HMAC_SHA256(HMAC_SECRET, "asn:" + uid + ":" + assignmentId)
- *
- * Stable within (uid, assignmentId), unlinkable across assignments, and
- * unlinkable to a student without both the HMAC secret AND the OneRoster
- * roster.
- */
-export const getAssignmentPseudonymV1 = onCall(
-  {
-    memory: '256MiB',
-    secrets: [STUDENT_PSEUDONYM_HMAC_SECRET],
-    invoker: 'public',
-    cors: ALLOWED_ORIGINS,
-  },
-  (request) => {
-    if (!request.auth) {
-      throw new HttpsError('unauthenticated', 'Sign in required.');
-    }
-    if (request.auth.token.studentRole !== true) {
-      throw new HttpsError('permission-denied', 'Student role required.');
-    }
-    const rawAssignmentId = (request.data as { assignmentId?: unknown })
-      ?.assignmentId;
-    const assignmentId =
-      typeof rawAssignmentId === 'string' ? rawAssignmentId : '';
-    if (!assignmentId || assignmentId.length > 200) {
-      throw new HttpsError('invalid-argument', 'Invalid assignmentId.');
-    }
-
-    const hmacSecret = STUDENT_PSEUDONYM_HMAC_SECRET.value();
-    if (!hmacSecret) {
-      throw new HttpsError('internal', 'Server configuration missing.');
-    }
-
-    const pseudonym = computeAssignmentPseudonym(
-      request.auth.uid,
-      assignmentId,
-      hmacSecret
-    );
-    return { pseudonym };
-  }
-);
-
-/**
- * getStudentClassDirectoryV1
- *
- * Returns class metadata (name, teacher display name, subject, code) for the
- * authenticated student's claim-bound `classIds`. Powers the sidebar on
- * `/my-assignments` so a student sees "English 9 / Ms. Halverson" instead of
- * an opaque sourcedId.
- *
- * Lookup order per classId:
- *   1. `collectionGroup('rosters').where('classlinkClassId', '==', classId)` —
- *      real ClassLink imports. Parent path gives teacher uid. Safe across
- *      orgs because ClassLink-issued sourcedIds are not admin-controlled.
- *   2. `organizations/{orgId}/testClasses/{classId}` — admin-managed test
- *      classes. Always read via the org-scoped doc path (never via a
- *      collectionGroup lookup on `testClassId`) because test class IDs are
- *      admin-chosen slugs that can collide across orgs; a collectionGroup
- *      query would risk returning another org's roster.
- *
- * PII: this function never returns student names, emails, or any field from
- * the per-roster Drive file. Only Firestore-side roster meta (which is
- * itself PII-free) plus the teacher's own `displayName` from Firebase Auth.
- * Teacher names are organizational data, not student PII.
- *
- * Failure: classIds that match nothing are silently dropped from the
- * response so the sidebar simply omits them. The page renders a fallback
- * label client-side rather than treating a missing entry as an error.
- */
-export const getStudentClassDirectoryV1 = onCall(
-  {
-    memory: '256MiB',
-    invoker: 'public',
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    if (!request.auth) {
-      throw new HttpsError('unauthenticated', 'Sign in required.');
-    }
-    if (request.auth.token.studentRole !== true) {
-      throw new HttpsError('permission-denied', 'Student role required.');
-    }
-
-    const rawClassIds: unknown = request.auth.token.classIds;
-    if (!Array.isArray(rawClassIds)) {
-      throw new HttpsError('failed-precondition', 'No classes on token.');
-    }
-    const classIds = rawClassIds
-      .filter((c): c is string => typeof c === 'string' && c.length > 0)
-      .slice(0, STUDENT_LOGIN_CLASS_IDS_MAX);
-    if (classIds.length === 0) {
-      return { classes: [] };
-    }
-
-    const orgId =
-      typeof request.auth.token.orgId === 'string'
-        ? request.auth.token.orgId
-        : '';
-
-    const db = admin.firestore();
-
-    // Per-call cache: many classes may share a teacher; one Auth lookup
-    // suffices.
-    const teacherNameCache = new Map<string, string>();
-    const resolveTeacherName = async (teacherUid: string): Promise<string> => {
-      const cached = teacherNameCache.get(teacherUid);
-      if (cached !== undefined) return cached;
-      try {
-        const user = await admin.auth().getUser(teacherUid);
-        const displayName =
-          (typeof user.displayName === 'string' && user.displayName) ||
-          (typeof user.email === 'string' ? user.email.split('@')[0] : '') ||
-          '';
-        teacherNameCache.set(teacherUid, displayName);
-        return displayName;
-      } catch {
-        // Auth lookup can fail for legacy / deleted teacher accounts. Caller
-        // falls back to an empty teacher name; the row still renders.
-        teacherNameCache.set(teacherUid, '');
-        return '';
-      }
-    };
-
-    interface DirectoryEntry {
-      classId: string;
-      name: string;
-      teacherDisplayName: string;
-      subject?: string;
-      code?: string;
-    }
-
-    // Batch the per-classId collectionGroup queries instead of fanning out
-    // a separate equality query per id. Firestore caps `in`-array size at
-    // 30; we chunk at 10 to stay safely under the limit and to fit the
-    // typical `STUDENT_LOGIN_CLASS_IDS_MAX = 20` payload in 2 chunks.
-    const FIRESTORE_IN_CHUNK_SIZE = 10;
-
-    /**
-     * Run `collectionGroup('rosters').where(field, 'in', chunk)` for every
-     * chunk of `ids`, then index the matching roster docs by their
-     * `field`-value. First match wins on duplicates so the teacher whose
-     * roster Firestore returns first deterministically owns the directory
-     * entry for that class.
-     */
-    const batchLookupByField = async (
-      field: 'classlinkClassId',
-      ids: readonly string[]
-    ): Promise<Map<string, FirebaseFirestore.QueryDocumentSnapshot>> => {
-      const out = new Map<string, FirebaseFirestore.QueryDocumentSnapshot>();
-      if (ids.length === 0) return out;
-      const snapshots = await Promise.all(
-        chunk(ids, FIRESTORE_IN_CHUNK_SIZE).map((idChunk) =>
-          db.collectionGroup('rosters').where(field, 'in', idChunk).get()
-        )
-      );
-      for (const snap of snapshots) {
-        for (const doc of snap.docs) {
-          const value: unknown = doc.get(field);
-          if (typeof value === 'string' && !out.has(value)) {
-            out.set(value, doc);
-          }
-        }
-      }
-      return out;
-    };
-
-    const buildEntryFromRoster = async (
-      classId: string,
-      rosterDoc: FirebaseFirestore.QueryDocumentSnapshot,
-      includeCode: boolean
-    ): Promise<DirectoryEntry> => {
-      const data = rosterDoc.data();
-      const teacherUid = rosterDoc.ref.parent.parent?.id ?? '';
-      const teacherDisplayName = teacherUid
-        ? await resolveTeacherName(teacherUid)
-        : '';
-      return {
-        classId,
-        name: typeof data.name === 'string' ? data.name : classId,
-        teacherDisplayName,
-        subject:
-          typeof data.classlinkSubject === 'string'
-            ? data.classlinkSubject
-            : undefined,
-        code:
-          includeCode && typeof data.classlinkClassCode === 'string'
-            ? data.classlinkClassCode
-            : undefined,
-      };
-    };
-
-    // 1. Batched collectionGroup lookup for ClassLink classes. Two `in`
-    // queries per field chunk replace what was up to 20 separate equality
-    // queries. Real ClassLink sourcedIds are issued globally by ClassLink
-    // and are not admin-controlled, so collisions across orgs are not a
-    // realistic risk on this branch.
-    const classlinkMatches = await batchLookupByField(
-      'classlinkClassId',
-      classIds
-    );
-    const unresolvedAfterClasslink = classIds.filter(
-      (id) => !classlinkMatches.has(id)
-    );
-
-    // 2. Direct testClasses doc reads — for every classId not resolved as
-    // ClassLink. We deliberately do NOT use a `collectionGroup('rosters')
-    // .where('testClassId', 'in', …)` lookup here: testClassIds are
-    // admin-chosen slugs (default `testclass`, or a slugified title) and
-    // can collide across orgs, so a collectionGroup query would risk
-    // returning another org's roster doc and leaking that teacher's name
-    // and class metadata to the student. The org-scoped document path is
-    // gated by the student's verified `orgId` claim, so it cannot cross
-    // org boundaries. The trade-off is that we no longer surface the
-    // importing teacher's display name on test-class directory entries —
-    // these are admin-managed mocks where teacher attribution is
-    // cosmetic.
-    const testClassDocs = new Map<string, FirebaseFirestore.DocumentSnapshot>();
-    if (orgId && unresolvedAfterClasslink.length > 0) {
-      const docs = await Promise.all(
-        unresolvedAfterClasslink.map((id) =>
-          db
-            .doc(`organizations/${orgId}/testClasses/${id}`)
-            .get()
-            .catch(() => null)
-        )
-      );
-      for (let i = 0; i < unresolvedAfterClasslink.length; i++) {
-        const d = docs[i];
-        if (d && d.exists) testClassDocs.set(unresolvedAfterClasslink[i], d);
-      }
-    }
-
-    const entries = await Promise.all(
-      classIds.map(async (classId): Promise<DirectoryEntry | null> => {
-        const fromClasslink = classlinkMatches.get(classId);
-        if (fromClasslink) {
-          return buildEntryFromRoster(classId, fromClasslink, true);
-        }
-        const fromTestClassDoc = testClassDocs.get(classId);
-        if (fromTestClassDoc) {
-          const data = fromTestClassDoc.data() ?? {};
-          return {
-            classId,
-            name:
-              typeof data.title === 'string' && data.title.length > 0
-                ? data.title
-                : classId,
-            teacherDisplayName: '',
-            subject:
-              typeof data.subject === 'string' ? data.subject : undefined,
-          };
-        }
-        return null;
-      })
-    );
-
-    const classes = entries.filter((e): e is DirectoryEntry => e !== null);
-    return { classes };
-  }
-);
-
-/**
- * getPseudonymsForAssignmentV1
- *
- * Called by a teacher's client when rendering the grading view for an
- * assignment. Returns both pseudonyms plus names for every student in the
- * targeted ClassLink class:
- *   { sourcedId -> { studentUid, assignmentPseudonym, givenName, familyName } }
- * so the teacher's client can join Firestore responses (keyed by either the
- * session-scoped studentUid for quiz/video/guided-learning or the
- * assignment-scoped pseudonym for mini-app submissions) back to roster
- * identity and display names. Names never touch Firestore — they stay in
- * teacher-browser memory for the session. The HMAC secret never leaves the
- * server.
- *
- * Only callable by a teacher who actually teaches the requested class
- * (ClassLink membership is re-verified on every call).
- */
-export const getPseudonymsForAssignmentV1 = onCall(
-  {
-    memory: '256MiB',
-    minInstances: 1,
-    cors: ALLOWED_ORIGINS,
-    secrets: [
-      CLASSLINK_CLIENT_ID,
-      CLASSLINK_CLIENT_SECRET,
-      CLASSLINK_TENANT_URL,
-      STUDENT_PSEUDONYM_HMAC_SECRET,
-    ],
-    invoker: 'public',
-  },
-  async (request) => {
-    if (!request.auth) {
-      throw new HttpsError('unauthenticated', 'Sign in required.');
-    }
-    // Teachers authenticate with standard Firebase Auth (email present on
-    // token). Students never have email on their token, so this also keeps
-    // students out of the teacher-only endpoint.
-    const teacherEmail = request.auth.token.email;
-    if (!teacherEmail || request.auth.token.studentRole === true) {
-      throw new HttpsError('permission-denied', 'Teacher account required.');
-    }
-
-    const data = request.data as {
-      assignmentId?: unknown;
-      classId?: unknown;
-      orgId?: unknown;
-    };
-    const assignmentId =
-      typeof data?.assignmentId === 'string' ? data.assignmentId : '';
-    const classId = typeof data?.classId === 'string' ? data.classId : '';
-    // orgId is optional for backwards compatibility (older clients). The
-    // test-class branch only activates when it's provided AND the requested
-    // classId resolves to a `testClasses` doc under that org.
-    const orgId = typeof data?.orgId === 'string' ? data.orgId : '';
-    if (!assignmentId || !classId) {
-      throw new HttpsError(
-        'invalid-argument',
-        'assignmentId and classId are required.'
-      );
-    }
-
-    const hmacSecret = STUDENT_PSEUDONYM_HMAC_SECRET.value();
-    const classlinkClientId = CLASSLINK_CLIENT_ID.value();
-    const classlinkClientSecret = CLASSLINK_CLIENT_SECRET.value();
-    const tenantUrl = CLASSLINK_TENANT_URL.value();
-    if (
-      !hmacSecret ||
-      !classlinkClientId ||
-      !classlinkClientSecret ||
-      !tenantUrl
-    ) {
-      throw new HttpsError('internal', 'Server configuration missing.');
-    }
-
-    // ── Test-class branch ────────────────────────────────────────────────
-    // Test classes (admin-managed mocks under `organizations/{orgId}/testClasses`)
-    // bypass ClassLink entirely. Their students log in via the `studentLoginV1`
-    // test bypass, which mints UIDs as `HMAC("sid:test:{emailLower}", secret)`.
-    // ClassLink OneRoster has no record of them, so the standard branch returns
-    // an empty pseudonym map and the teacher monitor falls back to "Student".
-    // This branch resolves names from the `memberEmails` array on the test-class
-    // doc, using the email local-part as the display name (matching what
-    // `materializeTestClassStudents` shows in the import dialog).
-    if (orgId) {
-      const db = admin.firestore();
-      const teacherEmailLower = teacherEmail.toLowerCase();
-      const memberRef = db.doc(
-        `organizations/${orgId}/members/${teacherEmailLower}`
-      );
-      const memberSnap = await memberRef.get();
-      if (!memberSnap.exists) {
-        throw new HttpsError(
-          'permission-denied',
-          'Not a member of this organization.'
-        );
-      }
-
-      const testClassRef = db.doc(
-        `organizations/${orgId}/testClasses/${classId}`
-      );
-      const testClassSnap = await testClassRef.get();
-      if (testClassSnap.exists) {
-        // Authorize: teacher must own a roster whose `testClassId` matches.
-        // Roster metadata lives in Firestore (no Drive read needed for this
-        // gate). Same trust model as the ClassLink branch's "teaches this
-        // class" check, but anchored to the teacher's own roster ownership.
-        const ownedRosters = await db
-          .collection(`users/${request.auth.uid}/rosters`)
-          .where('testClassId', '==', classId)
-          .limit(1)
-          .get();
-        if (ownedRosters.empty) {
-          throw new HttpsError(
-            'permission-denied',
-            'Not a teacher of this test class.'
-          );
-        }
-
-        const testClassData = (testClassSnap.data() ?? {}) as {
-          memberEmails?: unknown;
-        };
-        const memberEmails = Array.isArray(testClassData.memberEmails)
-          ? testClassData.memberEmails.filter(
-              (e): e is string => typeof e === 'string' && e.length > 0
-            )
-          : [];
-
-        const pseudonyms: Record<
-          string,
-          {
-            studentUid: string;
-            assignmentPseudonym: string;
-            givenName: string;
-            familyName: string;
-          }
-        > = {};
-        for (const rawEmail of memberEmails) {
-          const emailLower = rawEmail.toLowerCase();
-          // Mirrors `studentLoginV1` test-bypass UID minting at
-          // functions/src/index.ts ~2868: HMAC over "sid:test:{emailLower}".
-          const studentUid = computeStudentUid(
-            `test:${emailLower}`,
-            hmacSecret
-          );
-          // Display name = email local-part. This matches what the import
-          // dialog already shows (`materializeTestClassStudents` line 66:
-          // `firstName: email.split('@')[0]`), so the monitor view stays
-          // consistent with the roster.
-          const localPart = emailLower.split('@')[0] || emailLower;
-          // Key by the lowercased email (no `sourcedId` exists for test
-          // students). The client-side hook only iterates `Object.values`
-          // and re-keys by `studentUid`, so the key choice is internal.
-          pseudonyms[emailLower] = {
-            studentUid,
-            assignmentPseudonym: computeAssignmentPseudonym(
-              studentUid,
-              assignmentId,
-              hmacSecret
-            ),
-            givenName: localPart,
-            familyName: '',
-          };
-        }
-        return { pseudonyms };
-      }
-    }
-
-    const cleanTenantUrl = tenantUrl.replace(/\/$/, '');
-
-    // Verify the teacher actually teaches this class before disclosing the
-    // roster pseudonyms. A teacher can only retrieve pseudonyms for their
-    // own classes.
-    try {
-      if (!isSafeEmailForOneRosterFilter(teacherEmail)) {
-        throw new HttpsError(
-          'not-found',
-          'Teacher not found in ClassLink roster.'
-        );
-      }
-      const teacherUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users`;
-      const teacherParams = { filter: `email='${teacherEmail}'` };
-      const teacherHeaders = getOAuthHeaders(
-        teacherUrl,
-        teacherParams,
-        'GET',
-        classlinkClientId,
-        classlinkClientSecret
-      );
-      const teacherResp = await axios.get<{ users: OneRosterUserWithRole[] }>(
-        teacherUrl,
-        { params: teacherParams, headers: { ...teacherHeaders } }
-      );
-      const teacherUser = (teacherResp.data.users ?? [])[0];
-      if (!teacherUser) {
-        throw new HttpsError(
-          'not-found',
-          'Teacher not found in ClassLink roster.'
-        );
-      }
-      const classesUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/users/${teacherUser.sourcedId}/classes`;
-      const classesHeaders = getOAuthHeaders(
-        classesUrl,
-        {},
-        'GET',
-        classlinkClientId,
-        classlinkClientSecret
-      );
-      const classesResp = await axios.get<{ classes: ClassLinkClass[] }>(
-        classesUrl,
-        { headers: { ...classesHeaders } }
-      );
-      const teaches = (classesResp.data.classes ?? []).some(
-        (c) => c.sourcedId === classId
-      );
-      if (!teaches) {
-        throw new HttpsError(
-          'permission-denied',
-          'Not a teacher of this class.'
-        );
-      }
-
-      // Now fetch the class's students and compute the pseudonym map.
-      const studentsUrl = `${cleanTenantUrl}${ONEROSTER_BASE}/classes/${classId}/students`;
-      const studentsHeaders = getOAuthHeaders(
-        studentsUrl,
-        {},
-        'GET',
-        classlinkClientId,
-        classlinkClientSecret
-      );
-      const studentsResp = await axios.get<{ users: ClassLinkStudent[] }>(
-        studentsUrl,
-        { headers: { ...studentsHeaders } }
-      );
-      const students = studentsResp.data.users ?? [];
-
-      // Return both pseudonyms so teacher viewers can match whichever one
-      // the response doc is keyed by:
-      //  - studentUid            — HMAC(sourcedId, secret); equals the
-      //                            ClassLink student's Firebase Auth UID.
-      //                            Used by quiz/video-activity/guided-learning
-      //                            response docs that key on auth.currentUser.uid.
-      //  - assignmentPseudonym   — HMAC(studentUid + assignmentId, secret).
-      //                            Used by mini-app submission docs that key
-      //                            on a per-assignment opaque id.
-      const pseudonyms: Record<
-        string,
-        {
-          studentUid: string;
-          assignmentPseudonym: string;
-          givenName: string;
-          familyName: string;
-        }
-      > = {};
-      for (const s of students) {
-        if (!s.sourcedId) continue;
-        const studentUid = computeStudentUid(s.sourcedId, hmacSecret);
-        pseudonyms[s.sourcedId] = {
-          studentUid,
-          assignmentPseudonym: computeAssignmentPseudonym(
-            studentUid,
-            assignmentId,
-            hmacSecret
-          ),
-          givenName: s.givenName ?? '',
-          familyName: s.familyName ?? '',
-        };
-      }
-      return { pseudonyms };
-    } catch (err) {
-      if (err instanceof HttpsError) throw err;
-      if (axios.isAxiosError(err)) {
-        console.error(
-          '[getPseudonymsForAssignmentV1] ClassLink request failed:',
-          err.response?.status
-        );
-      } else {
-        console.error('[getPseudonymsForAssignmentV1] Unexpected failure.');
-      }
-      throw new HttpsError('internal', 'Roster service unavailable.');
-    }
-  }
-);
-
-// ─── PIN → SSO identity unification (Phase 3) ────────────────────────────────
-//
-// `pinLoginV1` lets a PIN-joining student authenticate with a custom token
-// whose uid is the same HMAC pseudonym `studentLoginV1` would mint for them
-// if they came in via SSO. Once they sign in with that token, the per-session
-// response doc keys by their `auth.uid` (= the SSO uid), so a student who
-// joins one launch via SSO and another via PIN converges on the same response
-// doc and the per-session attempt cap holds across both auth paths.
-//
-// `commitRosterPinIndexV1` is the teacher-side companion: when a teacher saves
-// a ClassLink-origin roster, the client posts the (period, pin, sourcedId)
-// tuples and this function writes the non-PII pin_index sidecar that
-// pinLoginV1 reads. PII (names, emails) never leaves Drive — the index holds
-// only opaque hashes and ids.
-
-const PIN_INDEX_SUBCOLLECTION = 'pin_index';
-const PIN_INDEX_MAX_ENTRIES = 200;
-
-/**
- * Mirror of `encodeResponseKeySegment` in `useQuizSession.ts`. Duplicated
- * server-side rather than imported because the functions package has its own
- * tsconfig + emit and the client hook lives outside it. The client + server
- * encoders MUST stay in lockstep — a divergence would produce mismatched
- * `indexKey`s and silently break the PIN→SSO lookup.
- */
-function encodeResponseKeySegment(value: string | undefined): string {
-  const trimmed = (value ?? '').trim();
-  if (!trimmed) return 'default';
-  const normalized = trimmed.toLowerCase().replace(/[^a-z0-9]+/g, '_');
-  const stripped = normalized.replace(/^_+|_+$/g, '');
-  return stripped || 'default';
-}
-
-function pinIndexKey(period: string, pin: string): string {
-  return `${encodeResponseKeySegment(period)}__${encodeResponseKeySegment(pin)}`;
-}
-
-interface CommitRosterPinIndexEntry {
-  period: string;
-  pin: string;
-  classlinkSourcedId: string;
-}
-
-/**
- * commitRosterPinIndexV1
- *
- * Input:
- *   {
- *     rosterId: string,
- *     entries: Array<{ period, pin, classlinkSourcedId }>
- *   }
- *
- * The caller MUST be the teacher who owns the roster (the function checks
- * `users/{auth.uid}/rosters/{rosterId}` exists). The full set of entries
- * replaces the existing pin_index for the roster — entries dropped from the
- * input list have their corresponding index docs deleted in the same batch,
- * so a student removed from the roster automatically loses their
- * pin_index entry.
- *
- * The function reads the roster doc to recover `classlinkClassId` (used as
- * the index entry's `classId`) and `classlinkOrgId` (used for telemetry —
- * not written into the index). Only ClassLink-origin rosters with a
- * `classlinkClassId` produce an index; legacy local rosters skip silently
- * (no error) so the client's "build the index after every save" hook is
- * idempotent across both kinds.
- */
-export const commitRosterPinIndexV1 = onCall(
-  {
-    memory: '256MiB',
-    secrets: [STUDENT_PSEUDONYM_HMAC_SECRET],
-    invoker: 'public',
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    if (!request.auth) {
-      throw new HttpsError('unauthenticated', 'Sign in required.');
-    }
-    // Students cannot rebuild a teacher's index. The client-side caller is
-    // always the teacher who saved the roster.
-    if (request.auth.token.studentRole === true) {
-      throw new HttpsError('permission-denied', 'Teacher role required.');
-    }
-
-    const rawData = (request.data ?? {}) as {
-      rosterId?: unknown;
-      entries?: unknown;
-    };
-    const rosterId =
-      typeof rawData.rosterId === 'string' ? rawData.rosterId : '';
-    if (!rosterId) {
-      throw new HttpsError('invalid-argument', 'rosterId is required.');
-    }
-    if (!Array.isArray(rawData.entries)) {
-      throw new HttpsError('invalid-argument', 'entries must be an array.');
-    }
-    if (rawData.entries.length > PIN_INDEX_MAX_ENTRIES) {
-      throw new HttpsError(
-        'invalid-argument',
-        `entries exceeds the max of ${PIN_INDEX_MAX_ENTRIES}.`
-      );
-    }
-
-    const entries: CommitRosterPinIndexEntry[] = [];
-    let skippedMalformed = 0;
-    for (const e of rawData.entries) {
-      if (typeof e !== 'object' || e === null) {
-        skippedMalformed++;
-        continue;
-      }
-      const period = (e as { period?: unknown }).period;
-      const pin = (e as { pin?: unknown }).pin;
-      const sourcedId = (e as { classlinkSourcedId?: unknown })
-        .classlinkSourcedId;
-      if (
-        typeof period !== 'string' ||
-        typeof pin !== 'string' ||
-        typeof sourcedId !== 'string' ||
-        period.length === 0 ||
-        pin.length === 0 ||
-        sourcedId.length === 0
-      ) {
-        // Skip malformed entries; don't fail the whole rebuild. Counted
-        // and surfaced in the response so the client can warn the
-        // teacher when one student's row is silently dropped (a typo'd
-        // ClassLink id otherwise produces an unindexed student who
-        // bypasses the cross-launch cap on legacy PIN).
-        skippedMalformed++;
-        continue;
-      }
-      entries.push({ period, pin, classlinkSourcedId: sourcedId });
-    }
-    if (skippedMalformed > 0) {
-      console.warn(
-        `[commitRosterPinIndexV1] Skipped ${skippedMalformed} malformed entries in roster ${rosterId}`
-      );
-    }
-
-    const hmacSecret = STUDENT_PSEUDONYM_HMAC_SECRET.value();
-    if (!hmacSecret) {
-      throw new HttpsError('internal', 'Server configuration missing.');
-    }
-
-    const db = admin.firestore();
-    const rosterRef = db
-      .collection('users')
-      .doc(request.auth.uid)
-      .collection('rosters')
-      .doc(rosterId);
-    const rosterSnap = await rosterRef.get();
-    if (!rosterSnap.exists) {
-      throw new HttpsError('not-found', 'Roster not found.');
-    }
-    const rosterData = rosterSnap.data() ?? {};
-    const classlinkClassId =
-      typeof rosterData.classlinkClassId === 'string' &&
-      rosterData.classlinkClassId.length > 0
-        ? rosterData.classlinkClassId
-        : null;
-    // Optional on the roster doc; empty-string default keeps the
-    // pin_index entry shape stable so `pinLoginV1` can read a string
-    // field unconditionally. Stored alongside `classId` so the login
-    // path doesn't need a `collectionGroup` lookup to recover orgId.
-    const classlinkOrgId =
-      typeof rosterData.classlinkOrgId === 'string'
-        ? rosterData.classlinkOrgId
-        : '';
-    if (!classlinkClassId) {
-      // Local rosters have no ClassLink class id and so can't bridge into
-      // the SSO uid space. Returning success (with `wrote: 0`) keeps the
-      // client's "save then commit index" hook idempotent across roster
-      // types — it doesn't have to special-case origin.
-      return {
-        wrote: 0,
-        deleted: 0,
-        skippedMalformed,
-        skippedReason: 'no-classlink-class-id' as const,
-      };
-    }
-
-    // Compute the desired set of (indexKey -> doc payload) tuples.
-    const desired = new Map<
-      string,
-      {
-        pseudonym: string;
-        classId: string;
-        orgId: string;
-        period: string;
-        updatedAt: number;
-      }
-    >();
-    const now = Date.now();
-    for (const entry of entries) {
-      const key = pinIndexKey(entry.period, entry.pin);
-      // Last-write-wins on intra-batch dupes (same encoded period+pin).
-      desired.set(key, {
-        pseudonym: computeStudentUid(entry.classlinkSourcedId, hmacSecret),
-        classId: classlinkClassId,
-        orgId: classlinkOrgId,
-        period: entry.period,
-        updatedAt: now,
-      });
-    }
-
-    // Read the current index so we know which docs to delete (entries that
-    // existed before but are not in `desired`). Bounded by
-    // PIN_INDEX_MAX_ENTRIES via `.limit()` — a roster with more entries
-    // would have been rejected on the input side already.
-    const pinIndexCollection = rosterRef.collection(PIN_INDEX_SUBCOLLECTION);
-    const existingSnap = await pinIndexCollection
-      .limit(PIN_INDEX_MAX_ENTRIES + 1)
-      .get();
-
-    const batch = db.batch();
-    let deleted = 0;
-    for (const docSnap of existingSnap.docs) {
-      if (!desired.has(docSnap.id)) {
-        batch.delete(docSnap.ref);
-        deleted++;
-      }
-    }
-
-    let wrote = 0;
-    for (const [key, payload] of desired) {
-      batch.set(pinIndexCollection.doc(key), payload);
-      wrote++;
-    }
-
-    await batch.commit();
-    return { wrote, deleted, skippedMalformed };
-  }
-);
-
-interface PinLoginRequestData {
-  kind?: unknown;
-  sessionId?: unknown;
-  code?: unknown;
-  pin?: unknown;
-  period?: unknown;
-}
-
-/**
- * pinLoginV1
- *
- * Input:
- *   {
- *     kind: 'quiz' | 'video-activity',
- *     sessionId?: string,    // required for video-activity
- *     code?: string,         // required for quiz
- *     pin: string,
- *     period?: string,       // optional, picks one when the session has
- *                            // multiple rosters and the pin is ambiguous
- *   }
- *
- * Resolves the (session, period, pin) tuple to a roster-bound student
- * identity by reading the teacher's pin_index sidecar (built by
- * `commitRosterPinIndexV1`). On success returns a custom token whose uid
- * matches the same HMAC pseudonym `studentLoginV1` would mint for that
- * student over SSO, plus `studentRole: true` and `classIds: [classId]`
- * so the student passes the response-rule class gate.
- *
- * On no-match returns `{ matched: false }` so the client can fall back
- * to the existing anonymous PIN flow (legacy non-rostered sessions, or
- * rosters whose index hasn't been built yet).
- *
- * No PII logging — only class-of-failure counters.
- */
-export const pinLoginV1 = onCall(
-  {
-    memory: '256MiB',
-    secrets: [STUDENT_PSEUDONYM_HMAC_SECRET],
-    invoker: 'public',
-    cors: ALLOWED_ORIGINS,
-  },
-  async (request) => {
-    const data = (request.data ?? {}) as PinLoginRequestData;
-    const kind =
-      data.kind === 'quiz' || data.kind === 'video-activity' ? data.kind : null;
-    const pin = typeof data.pin === 'string' ? data.pin.trim() : '';
-    const period = typeof data.period === 'string' ? data.period : '';
-    if (!kind) {
-      throw new HttpsError('invalid-argument', 'kind is required.');
-    }
-    if (!pin) {
-      throw new HttpsError('invalid-argument', 'pin is required.');
-    }
-
-    const db = admin.firestore();
-
-    // Resolve the session.
-    let sessionRef: admin.firestore.DocumentReference;
-    if (kind === 'quiz') {
-      const code =
-        typeof data.code === 'string' ? normalizeQuizCode(data.code) : '';
-      if (!code) {
-        throw new HttpsError('invalid-argument', 'code is required for quiz.');
-      }
-      const codeMatch = await db
-        .collection('quiz_sessions')
-        .where('code', '==', code)
-        .limit(5)
-        .get();
-      const joinable = codeMatch.docs.find((d) => {
-        const s = (d.data() as { status?: unknown }).status;
-        return s === 'waiting' || s === 'active' || s === 'paused';
-      });
-      if (!joinable) {
-        console.warn('[pinLoginV1] fallback', {
-          kind,
-          reason: 'no-joinable-session',
-          codeMatchCount: codeMatch.size,
-          period,
-        });
-        return { matched: false, reason: 'no-joinable-session' };
-      }
-      sessionRef = joinable.ref;
-    } else {
-      const sessionId =
-        typeof data.sessionId === 'string' ? data.sessionId : '';
-      if (!sessionId) {
-        throw new HttpsError(
-          'invalid-argument',
-          'sessionId is required for video-activity.'
-        );
-      }
-      sessionRef = db.collection('video_activity_sessions').doc(sessionId);
-    }
-
-    const sessionSnap = await sessionRef.get();
-    if (!sessionSnap.exists) {
-      console.warn('[pinLoginV1] fallback', {
-        kind,
-        reason: 'session-not-found',
-        sessionId: sessionRef.id,
-        period,
-      });
-      return { matched: false, reason: 'session-not-found' };
-    }
-    const sessionData = sessionSnap.data() ?? {};
-    const teacherUid =
-      typeof sessionData.teacherUid === 'string' ? sessionData.teacherUid : '';
-    if (!teacherUid) {
-      console.warn('[pinLoginV1] fallback', {
-        kind,
-        reason: 'session-missing-teacher',
-        sessionId: sessionRef.id,
-        period,
-      });
-      return { matched: false, reason: 'session-missing-teacher' };
-    }
-    const rosterIds = Array.isArray(sessionData.rosterIds)
-      ? sessionData.rosterIds.filter(
-          (r: unknown): r is string => typeof r === 'string' && r.length > 0
-        )
-      : [];
-    if (rosterIds.length === 0) {
-      // No roster on the session — can't bridge. Fall through to the
-      // legacy anonymous PIN flow on the client. Common for legacy
-      // PIN-only sessions; surface so we can spot a rostered session
-      // that lost its rosterIds via a bad write.
-      console.warn('[pinLoginV1] fallback', {
-        kind,
-        reason: 'no-rosters-on-session',
-        sessionId: sessionRef.id,
-        teacherUid,
-        period,
-      });
-      return { matched: false, reason: 'no-rosters-on-session' };
-    }
-
-    const indexKey = pinIndexKey(period, pin);
-
-    // Probe each roster's pin_index for the indexKey IN PARALLEL.
-    // PIN-bridged join is on the hot path of every PIN-joining student,
-    // and a multi-class session (multiple rosters) would otherwise
-    // serialize one .get() per roster. Fan-out is bounded by
-    // STUDENT_LOGIN_CLASS_IDS_MAX-style sizing on the client side
-    // (rosterIds is teacher-authored and small in practice).
-    //
-    // A multi-class session will typically have only one match because
-    // PIN + encoded-period uniquely identifies a student. If multiple
-    // rosters happen to match (PIN collision across periods with a
-    // missing/default period), prefer the first hit by rosterIds order
-    // — same behavior the legacy PIN response-key resolution documents
-    // at `quizScoreboard.ts` `resolvePinName`.
-    //
-    // `orgId` lives directly on the index entry (Phase 3 review fix),
-    // so the login path is a single doc read per roster instead of an
-    // additional `collectionGroup('rosters').where('classlinkClassId'…)`
-    // scan to recover org metadata.
-    const indexRefs = rosterIds.map((rosterId) =>
-      db
-        .collection('users')
-        .doc(teacherUid)
-        .collection('rosters')
-        .doc(rosterId)
-        .collection(PIN_INDEX_SUBCOLLECTION)
-        .doc(indexKey)
-    );
-    const indexSnaps = await Promise.all(indexRefs.map((ref) => ref.get()));
-
-    let matched: {
-      pseudonym: string;
-      classId: string;
-      orgId: string;
-    } | null = null;
-    for (const indexSnap of indexSnaps) {
-      if (!indexSnap.exists) continue;
-      const entry = indexSnap.data() ?? {};
-      const pseudonym =
-        typeof entry.pseudonym === 'string' ? entry.pseudonym : '';
-      const classId = typeof entry.classId === 'string' ? entry.classId : '';
-      const orgId = typeof entry.orgId === 'string' ? entry.orgId : '';
-      if (pseudonym && classId) {
-        matched = { pseudonym, classId, orgId };
-        break;
-      }
-    }
-
-    if (!matched) {
-      // The PIN+period tuple didn't resolve to any roster entry across
-      // the session's rosters. This is the most diagnostically valuable
-      // fallback — it usually means either (a) the teacher hasn't run
-      // commitRosterPinIndexV1 since the roster was last edited, or
-      // (b) the period the student entered doesn't match the roster's
-      // encoded period. Period is logged (not the PIN itself) so we can
-      // diff against the roster's period encoding.
-      console.warn('[pinLoginV1] fallback', {
-        kind,
-        reason: 'no-index-entry',
-        sessionId: sessionRef.id,
-        teacherUid,
-        rosterCount: rosterIds.length,
-        period,
-      });
-      return { matched: false, reason: 'no-index-entry' };
-    }
-
-    let customToken: string;
-    try {
-      customToken = await admin.auth().createCustomToken(matched.pseudonym, {
-        studentRole: true,
-        orgId: matched.orgId,
-        classIds: [matched.classId],
-      });
-    } catch (err) {
-      console.error('[pinLoginV1] createCustomToken failed:', err);
-      throw new HttpsError('internal', 'Failed to mint auth token.');
-    }
-
-    return { matched: true, customToken };
-  }
-);
 
 // SPIKE — Google Classroom Add-on de-risk slice (student handshake + teacher
 // discovery attachment-create). Defined in their own module to keep this file
